% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ref.R
\name{add_config}
\alias{add_config}
\title{Add a config section to a vega plot}
\usage{
add_config(vis, ...)
}
\arguments{
\item{vis}{a ggvis object}

\item{...}{vega title properties}
}
\value{
a ggvis object
}
\description{
Add a config section to a vega plot
}
