% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvis.R
\name{add_data}
\alias{add_data}
\title{Add dataset to a visualisation}
\usage{
add_data(vis, data, name = deparse2(substitute(data)), add_suffix = TRUE)
}
\arguments{
\item{vis}{Visualisation to modify.}

\item{data}{Data set to add.}

\item{name}{Data of data - optional, but helps produce informative
error messages.}

\item{add_suffix}{Should a unique suffix be added to the data object's ID?
This should only be FALSE when the spec requires a data set with a
 specific name.}
}
\description{
Add dataset to a visualisation
}
\examples{
mtcars \%>\% ggvis(~mpg, ~wt) \%>\% layer_points()
NULL \%>\% ggvis(~mpg, ~wt) \%>\% add_data(mtcars) \%>\% layer_points()
}
