% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ref.R
\name{add_group_mark}
\alias{add_group_mark}
\title{Add a group mark}
\usage{
add_group_mark(group_mark, ...)
}
\arguments{
\item{group_mark}{a mark object object}

\item{...}{vega mark properties}
}
\value{
a ggvis object
}
\description{
Add a group mark
}
