% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_axis.R
\name{add_guide_axis}
\alias{add_guide_axis}
\title{Defunct function for adding an axis}
\usage{
add_guide_axis(...)
}
\arguments{
\item{...}{Other arguments.}
}
\description{
This function has been replaced with \code{\link{add_axis}}.
}
