% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guide_legend.R
\name{add_guide_legend}
\alias{add_guide_legend}
\title{Defunct function for adding a legend}
\usage{
add_guide_legend(...)
}
\arguments{
\item{...}{Other arguments.}
}
\description{
This function has been replaced with \code{\link{add_legend}}.
}

