% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ref.R
\name{add_legend_}
\alias{add_legend_}
\title{Add a vega legend to a vega vis}
\usage{
add_legend_(vis, ...)
}
\arguments{
\item{vis}{a ggvis object}

\item{...}{vega legend properties}
}
\value{
a ggvis object
}
\description{
Note: legend is simply appended without trying to identify other legends for the scale the legend applies to
}
