% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ref.R
\name{add_mark_}
\alias{add_mark_}
\title{Add a mark}
\usage{
add_mark_(vis, ...)
}
\arguments{
\item{vis}{a ggvis object}

\item{...}{vega title properties}
}
\value{
a ggvis object
}
\description{
Add a mark
}
