% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ref.R
\name{add_projection}
\alias{add_projection}
\title{Add a vega projection to a vega vis}
\usage{
add_projection(vis, ...)
}
\arguments{
\item{vis}{a ggvis object}

\item{...}{vega projection properties}
}
\value{
a ggvis object
}
\description{
Add a vega projection to a vega vis
}
