% Generated by roxygen2 (4.0.0): do not edit by hand
\name{add_props}
\alias{add_props}
\title{Add visual properties to a visualisation}
\usage{
add_props(vis, ..., .props = NULL, inherit = NULL, env = parent.frame())
}
\arguments{
\item{vis}{Visualisation to modify.}

\item{...}{Named visual properties.}
}
\description{
Add visual properties to a visualisation
}
\examples{
mtcars \%>\% ggvis(~wt, ~mpg) \%>\% layer_points()
mtcars \%>\% ggvis() \%>\% add_props(~wt, ~mpg) \%>\% layer_points()
mtcars \%>\% ggvis(~wt) \%>\% add_props(y = ~mpg) \%>\% layer_points()
}

