% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_relative_scales}
\alias{add_relative_scales}
\title{Add x_rel and y_rel scales}
\usage{
add_relative_scales(vis)
}
\arguments{
\item{vis}{A ggvis object.}
}
\description{
This function adds scales named \code{x_rel} and \code{y_rel}, each of which
has a domain of 0 to 1, and the range is the plot's width or height.
These scales are useful for positioning visual elements relative to the
plotting area. For example, with legends.
}
\seealso{
\code{\link{add_legend}} for a usage example.
}

