% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_scale}
\alias{add_scale}
\title{Add arbitrary scales to ggvis.}
\usage{
add_scale(vis, scale)
}
\arguments{
\item{vis}{Visualisation to modify.}

\item{scale}{Scale object}

\item{domain}{Either a vector with static values for the domain, or
a reactive that returns a such a vector.}
}
\description{
Add arbitrary scales to ggvis.
}
\keyword{internal}

