% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ref.R
\name{add_scale_}
\alias{add_scale_}
\title{Add a scale}
\usage{
add_scale_(vis, ...)
}
\arguments{
\item{vis}{a ggvis object}

\item{...}{vega scale properties}
}
\value{
a ggvis object
}
\description{
Add a scale
}
