% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ref.R
\name{add_signal_}
\alias{add_signal_}
\title{Add a vega signal to a vega vis}
\usage{
add_signal_(vis, ...)
}
\arguments{
\item{vis}{a ggvis object}

\item{...}{vega signal properties}
}
\value{
a ggvis object
}
\description{
Add a vega signal to a vega vis
}
