% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_axis.R
\name{add_title}
\alias{add_title}
\title{Add a title as an axis}
\usage{
add_title(vis, ..., title = "Plot Title")
}
\arguments{
\item{vis}{a ggvis plot}

\item{...}{parameters to be passed to the axis}

\item{title}{the title to apply}
}
\value{
a ggvis plot
}
\description{
Based on
https://stackoverflow.com/questions/25018598/add-a-plot-title-to-ggvis
}
\details{
BUG: If no x axis exists, this will replace the xaxis
}
\examples{
mtcars \%>\%
  dplyr::group_by(cyl) \%>\%
  dplyr::summarize(mpg = median(mpg)) \%>\%
  dplyr::mutate(cyl = as.factor(cyl)) \%>\%
  ggvis::ggvis(x=~cyl, y=~mpg) \%>\%
  ggvis::layer_bars() \%>\%
  ggvis::add_title(title="MPG by Cylinder")
}
