\name{as.vega}
\alias{as.vega}
\title{Coerce an gigvis object to a vega list.}
\usage{
  as.vega(x, ...)
}
\arguments{
  \item{x}{an object to convert to vega}
}
\value{
  a list. When converted to JSON, will be the type of
  structure that vega expects.
}
\description{
  This generic function powers the coercion of gigvis
  objects to vega compatible data structures.
}

