% Generated by roxygen2 (4.0.1): do not edit by hand
\name{auto_group}
\alias{auto_group}
\title{Automatically split data by groups}
\usage{
auto_group(vis)
}
\arguments{
\item{vis}{The ggvis visualisation to modify.}
}
\description{
Use \code{auto_split} to split up a dataset on all categorical variables
specified by props, and have each piece rendered by the same mark.
}
\examples{
# Make cyl a factor (as it really should be)
mtcars2 <- mtcars
mtcars2$cyl <- factor(mtcars2$cyl)

# One line
mtcars2 \%>\% ggvis(~disp,  ~mpg, stroke = ~cyl) \%>\% layer_paths()
# One line for each level of cyl
mtcars2 \%>\% ggvis(~disp,  ~mpg, stroke = ~cyl) \%>\% group_by(cyl) \%>\%
  layer_paths()
mtcars2 \%>\% ggvis(~disp,  ~mpg, stroke = ~cyl) \%>\% auto_group() \%>\%
  layer_paths()
}
\seealso{
To manually specify grouping variables, see \code{\link{group_by}}.
}

