% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bin_vector}
\alias{bin_vector}
\alias{bin_vector.POSIXct}
\alias{bin_vector.numeric}
\title{Bin vectors}
\usage{
bin_vector(x, weight = NULL, ...)

\method{bin_vector}{numeric}(x, weight = NULL, ..., width = NULL,
  center = NULL, boundary = NULL, right = TRUE, pad = TRUE)

\method{bin_vector}{POSIXct}(x, weight = NULL, ..., width = NULL,
  center = NULL, boundary = NULL, right = TRUE, pad = TRUE)
}
\arguments{
\item{x}{a vector to bin}

\item{weight}{if specified, an integer vector of the same length as \code{x}
representing the number of occurances of each value in \code{x}}

\item{...}{additional arguments passed through to instances of the generic}

\item{width}{the width of a bin}

\item{center}{the center of a bin}

\item{boundary}{the boundary of a bin.  \code{center} and \code{boundary} should
not both be specified.}

\item{right}{a logical indicating whether the right boundary of a bin is
included with the bin.}

\item{pad}{a logical indicatign whether the bins should be padded to include
an empty bin on each side.}
}
\description{
A generic and several implementations for binning vectors.
}

