% Generated by roxygen2 (4.0.0): do not edit by hand
\name{branch_f}
\alias{branch_f}
\title{Create a branch, given a function}
\usage{
branch_f(vis, fun)
}
\arguments{
\item{vis}{The ggvis visualisation to modify.}

\item{fun}{A function that takes a single argument, the current
visualisation as input, and returns a modified visualisation.}
}
\description{
Create a branch, given a function
}

