\docType{class}
\name{Click-class}
\alias{Click-class}
\title{Event broker for click events.}
\description{
The click event broker is useful if you want your shiny app
to respond to click events (\code{mouse_click} in a custom
way.
}
\seealso{
\code{\link{click_tooltip}} for a custom wrapper that can
use click events to display tooltips.
}

