% Generated by roxygen2 (4.0.0): do not edit by hand
\name{compute_width}
\alias{compute_width}
\title{Add "width" of data}
\usage{
compute_width(x, x_var, width = NULL)
}
\arguments{
\item{x}{Dataset-like object to count. Built-in methods for data frames,
grouped data frames and ggvis visualisations.}

\item{x_var}{Name of variable to compute width of.}

\item{width}{An absolute width to use. If \code{NULL} (the default), the
width will be equivalent to the resolution of the data.}
}
\value{
The original data frame, with additional columns:
 \item{xmin_}{left boundary of bin}
 \item{xmax_}{right boundary of bin}
 \item{width_}{width of bin}

The width of each "bin" is set to the resolution of the data -- that is, the
smallest difference between two x values.
}
\description{
This compute function is generally used in conjunction with
\code{\link{compute_count}}, when used on data with a continuous x variable.
By default, the computed width will be equal to the resolution of the data,
or, in other words the smallest difference between two values in the data.
}
\details{
An absolute width for each x can be specified by using the \code{width}
argument. If \code{width} is NULL (the default), it will use the resolution
of the data as the width.
}
\examples{
mtcars \%>\% compute_count(~disp) \%>\% compute_width(~x_)
mtcars \%>\% compute_count(~mpg) \%>\% compute_width(~x_)

# Use a specific width
pressure \%>\% compute_count(~temperature) \%>\% compute_width(~x_)
pressure \%>\% compute_count(~temperature) \%>\% compute_width(~x_, width = 5)

# It doesn't matter whether you transform inside or outside of a vis
mtcars \%>\% compute_count(~cyl, ~wt) \%>\%
  compute_width(~x_, width = .5) \%>\%
  ggvis(x = ~xmin_, x2 = ~xmax_, y = ~count_, y2 = 0) \%>\%
  layer_rects() \%>\%
  set_dscale("y", "numeric", domain = c(0, NA))

mtcars \%>\%
  ggvis(x = ~xmin_, x2 = ~xmax_, y = ~count_, y2 = 0) \%>\%
  compute_count(~cyl, ~wt) \%>\%
  compute_width(~x_) \%>\%
  layer_rects() \%>\%
  set_dscale("y", "numeric", domain = c(0, NA))
}
\seealso{
\code{\link{compute_bin}} For counting cases within ranges of
  a continuous variable.

\code{\link{compute_count}} For counting cases at specific values
  of a variable.
}

