% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dplyr-ggvis}
\alias{arrange.ggvis}
\alias{dplyr-ggvis}
\alias{filter.ggvis}
\alias{groups.ggvis}
\alias{mutate.ggvis}
\alias{regroup.ggvis}
\alias{select.ggvis}
\alias{summarise.ggvis}
\alias{ungroup.ggvis}
\title{Dplyr verbs for ggvis.}
\usage{
groups.ggvis(x)

regroup.ggvis(x, value)

ungroup.ggvis(x)

summarise.ggvis(.data, ...)

mutate.ggvis(.data, ...)

arrange.ggvis(.data, ...)

select.ggvis(.data, ...)

\method{filter}{ggvis}(.data, ...)
}
\description{
Dplyr verbs for ggvis.
}
\examples{
library(dplyr)
base <- mtcars \%>\% ggvis(~mpg, ~cyl) \%>\% layer_points()
base \%>\% group_by(cyl) \%>\% summarise(mpg = mean(mpg)) \%>\%
  layer_points(fill := "red", size := 100)

base \%>\% filter(mpg > 25) \%>\% layer_points(fill := "red")

base \%>\% mutate(cyl = jitter(cyl)) \%>\% layer_points(fill := "red")
}
\keyword{internal}

