% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvis.R
\name{dump_spec}
\alias{dump_spec}
\title{Export the ggvis vega schema json}
\usage{
dump_spec(x, ...)
}
\arguments{
\item{x}{a ggvis object}

\item{...}{other arguments passed to \code{as.vega}}
}
\description{
Export the ggvis vega schema json
}
