% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_.R
\name{dump_spec_}
\alias{dump_spec_}
\title{Export the ggvis vega schema json}
\usage{
dump_spec_(vis)
}
\arguments{
\item{vis}{a vega visualisation}
}
\value{
the vega spec
}
\description{
Export the ggvis vega schema json
}
