% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_encode.R
\name{encode_prop}
\alias{encode_prop}
\title{convert a line of a properties dataframe to a list of encode parameters}
\usage{
encode_prop(prop, scale = NULL)
}
\arguments{
\item{prop}{a one-line properties data frame. (If more lines exist, only the first
line will be used.)}

\item{scale}{string a scale name}
}
\value{
list of encode properties
}
\description{
convert a line of a properties dataframe to a list of encode parameters
}
