% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_is.R
\name{error_helper}
\alias{error_helper}
\title{parse the errors and raise them if they exist}
\usage{
error_helper(errors)
}
\arguments{
\item{errors}{dataframe of jsonvalidate errors}
}
\value{
Null (if not stopped)
}
\description{
from jsonvalidate::json_validator
}
