% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain.ggvis}
\alias{explain.ggvis}
\title{Print out the structure of a ggvis object in a friendly format}
\usage{
\method{explain}{ggvis}(x, ...)
}
\arguments{
\item{x}{Visualisation to explain}

\item{...}{Needed for compatibility with generic. Ignored by this method.}
}
\description{
Print out the structure of a ggvis object in a friendly format
}
