% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filter}
\alias{filter}
\alias{filter_}
\title{Filter}
\arguments{
\item{.data}{A tbl.}

\item{...}{Logical predicates. Multiple conditions are combined with &.}

\item{.dots}{Used to work around non-standard evaluation. See
vignette("nse", package="dplyr") for details.}
}
\description{
This is the same as \code{dplyr::\link[dplyr]{filter}} function.
See \code{dplyr::\link[dplyr]{filter}} for more information.
}

