% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip the x and y axis}
\usage{
flip(vis)
}
\arguments{
\item{vis}{a ggvis object}
}
\value{
a ggvis object
}
\description{
This is accomplished by updating the x & y marks, updating the flipping the
scales, and updating the axis labels.
}
\details{
WARNING: This currently works for rectangular layer figures.  It may not work with
multiple-layer figures, other marks, or signals.

WARNING: No tests currently exist for this function
}
\examples{
p <- mtcars \%>\%
dplyr::group_by(cyl) \%>\%
  dplyr::summarize(mpg = median(mpg)) \%>\%
  ggvis::ggvis(x=~cyl, y=~mpg) \%>\%
  ggvis::layer_bars() \%>\%
  ggvis::add_axis("x", title="Cylinders") \%>\%
  ggvis::add_axis("y", title="Miles Per Gallon") \%>\%
  ggvis::flip()
p <- mtcars \%>\%
  dplyr::group_by(cyl) \%>\%
  dplyr::summarize(mpg = median(mpg)) \%>\%
  ggvis::ggvis(x=~cyl, y=~mpg) \%>\%
  ggvis::layer_bars() \%>\%
  ggvis::flip()
}
