% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arc.R
\name{geom_arc}
\alias{geom_arc}
\title{Arcs}
\usage{
geom_arc(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., arrow = NULL, arrow.fill = NULL,
  lineend = "butt", linejoin = "round", res = pi/20, reverse = T,
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{arrow}{specification for arrow heads, as created by arrow().}

\item{arrow.fill}{fill colour to use for the arrow head (if closed). `NULL`
means use `colour` aesthetic.}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}
}
\description{
`geom_arc` draws an arc of center (x, y), with a custom radius, angle and width.
Couyld be useful for custom data visualisation,
or distribution or vector fields. For those scenarios it's more convenient than geom-segment.
}
\examples{

iris \%>\% ggplot(aes(Sepal.Length , Petal.Length, colour = Species)) + geom_arc(width = 0.8, aes(angle = Petal.Width ,radius = 0.1*Sepal.Width)) + coord_fixed()
}
