% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_data}
\alias{get_data}
\title{Get data from a ggvis object}
\usage{
get_data(vis)
}
\arguments{
\item{vis}{A ggvis object.}
}
\description{
This function is useful for inspecting the data in a ggvis object.
}
\examples{
p <- cocaine \%>\% ggvis(~price) \%>\% layer_bars()
get_data(p)
}

