% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_.R
\name{ggvega}
\alias{ggvega}
\title{Visualise a data set with a ggvis graphic.}
\usage{
ggvega(data = NULL, ...,
  schema = "https://vega.github.io/schema/vega/v3.0.json",
  description = NULL, background = NULL, width = NULL, height = NULL,
  padding = NULL, autosize = "pad", title = NULL)
}
\arguments{
\item{data}{A data object.}

\item{...}{Property mappings. If not named, the first two mappings are
taken to be \code{x} and \code{y}. Common properties are \code{x},
\code{y}, \code{stroke} (color), \code{fill}, \code{opacity}, \code{shape}}

\item{schema}{URL The URL for the Vega schema.}

\item{description}{String  A text description of the visualization.}

\item{background}{Color The background color of the entire view (defaults to transparent).}

\item{width}{Number  The width in pixels of the data rectangle.}

\item{height}{Number  The height in pixels of the data rectangle.}

\item{padding}{Number | Object The padding in pixels to add around the visualization. If a number, specifies padding for all sides. If an object, the value should have the format {"left": 5, "top": 5, "right": 5, "bottom": 5}. Padding is applied after autosize layout completes.}

\item{autosize}{String | Autosize Sets how the visualization size should be determined. If a string, should be one of pad (default), fit or none. Object values can additionally specify parameters for content sizing and automatic resizing. See the autosize section below for more.}

\item{title}{Title Title text to describe a visualization.}
}
\value{
visualization object
}
\description{
https://vega.github.io/vega/docs/specification/
}
