\name{+.ggvis}
\alias{+.ggvis}
\title{Modify a ggvis object by adding new components}
\usage{
\method{+}{ggvis}(e1, e2)
}
\arguments{
  \item{e1}{A ggvis object.}

  \item{e2}{Another object, of a type listed above.}
}
\description{
This operator lets you add objects to a ggvis object.
}
\details{
You can add the following types of objects to a ggvis
object: \itemize{ \item \code{mark} \item \code{branch}
\item \code{scale} \item \code{guide_legend} \item
\code{guide_axis} \item \code{props} \item \code{opts} }
}

