% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ggvis_dependencies}
\alias{ggvis_dependencies}
\title{HTML dependencies of a ggvis plot}
\usage{
ggvis_dependencies(absolute = FALSE, in_shiny = FALSE, dynamic = TRUE)
}
\arguments{
\item{absolute}{Use absolute paths? Needed for local html files.}

\item{in_shiny}{Will the plot be embedded in a shiny app? Drops jquery
dependency and adds shiny resource paths}

\item{dynamic}{Is this a dynamic plot? If so, add shiny-ggvis dependency}
}
\description{
There are five way to display a ggvis plot:
\itemize{
  \item static html file
  \item temporary shiny app
  \item static html embedded in Rmarkdown
  \item shiny app embedded in Rmarkdown
  \item embedded in regular shiny app
}
This function ensures that the right dependencies are specified for
each scenario.
}
\keyword{internal}

