% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggvis_message}
\alias{ggvis_message}
\title{Send a message to ggvis running on client}
\usage{
ggvis_message(session, type, data = NULL, id = NULL)
}
\arguments{
\item{session}{A session object.}

\item{type}{A string representing the type of the message.}

\item{data}{An object (typically a list) containing information for the client.}

\item{id}{A unique identifier for ggvis message handler (optional).}
}
\description{
This will be sent to the client and passed to a handler in ggvis.messages on
the client side. The handler is specified by \code{type}.
}

