% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggvisOutputElements}
\alias{ggvisOutputElements}
\title{Create HTML elements for ggvis output}
\usage{
ggvisOutputElements(plot_id = rand_id("plot_id"), spec = NULL,
  shiny = TRUE)
}
\arguments{
\item{plot_id}{Unique identifier to use for the div containing the ggvis plot.}

\item{spec}{Plot specification, used internally.}

\item{shiny}{Should this include headers for Shiny? For dynamic and
interactive plots, this should be TRUE; otherwise FALSE.}
}
\description{
This is an internal-facing function similar to ggvisOutput, but with more
options.
}
\keyword{internal}

