\name{gigvis_prerender}
\alias{gigvis_prerender}
\title{Given a filled gigvis spec object, output a gigvis_prerender object}
\usage{
  gigvis_prerender(gv, width = 600, height = 400,
    padding = c(20, 20, 30, 50), envir = parent.frame())
}
\description{
  Given a filled gigvis spec object, output a
  gigvis_prerender object
}

