% Generated by roxygen2 (4.0.2): do not edit by hand
\name{group_by}
\alias{group_by}
\title{Divide data into groups.}
\arguments{
\item{x}{a visualisation}

\item{...}{variables to group by.}

\item{add}{By default, when \code{add = FALSE}, \code{group_by} will
override existing groups. To instead add to the existing groups,
use \code{add = FALSE}}
}
\description{
Divide data into groups.
}

