% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prop_group.R
\name{groupwise}
\alias{groupwise}
\alias{is.prop_group}
\title{Groupwise property}
\usage{
groupwise(prop, offset = NULL, mult = NULL)

is.prop_group(x)
}
\arguments{
\item{prop}{Name of the group property. Typically "height" or "width".}

\item{offset,mult}{Additive and multiplicate offsets.}

\item{x}{object to test for group-ness}
}
\description{
\code{groupwise()} is used to refer to properties of the plot or
subvisualisation, typically the height or the width.
}
\examples{
# In the following example, we draw a rectangle running vertically
# across the plot by referring to its total height
ggvis(faithful) \%>\%
  layer_histograms(~eruptions) \%>\%
  emit_rects(props(
    x = ~mean(eruptions), width := 40,
    y := 0, y2 := groupwise("height"),
    fill := "red"
  ))

# The mult factor is useful to draw elements at fixed fractions
# of the plot:
ggvis(mtcars, ~mpg) \%>\%
  layer_histograms() \%>\%
  layer_axial_lines(y := groupwise("height", mult = 0.5),
    stroke := "red", strokeWidth := 4)
}

