\name{guess}
\alias{guess}
\alias{is.guess}
\title{Guess the value.}
\usage{
  guess()

  is.guess(x)
}
\arguments{
  \item{x}{object to test for guess-ness}
}
\description{
  This is an object used to indicate that the default value
  is to guess at the paramter. Guess values will always
  generate a message on the console indicating what the
  guess was. To supress this message supply a concrete
  value.
}

