% Generated by roxygen2 (4.0.0): do not edit by hand
\name{handle_brush}
\alias{handle_brush}
\title{Handle brush events on a visualisation.}
\usage{
handle_brush(vis, on_move = NULL)
}
\arguments{
\item{vis}{Visualisation to listen to.}

\item{on_move}{Callback function called with arguments \code{value} and
\code{session} every time the brush moves. Value is a list of points
under the brush.}
}
\description{
Handle brush events on a visualisation.
}
\examples{
# Display tooltip when objects are brushed
mtcars \%>\% ggvis(x = ~wt, y = ~mpg, size.brush := 400) \%>\%
  layer_points() \%>\%
  handle_brush(function(value, session) {
    show_tooltip(session, pagex = value$pagex2 + 5,
     pagey = value$pagey1 + 5, html = length(value$items))
  })
}

