% Generated by roxygen2 (4.0.0): do not edit by hand
\name{handle_click}
\alias{handle_click}
\alias{handle_hover}
\title{Handle mouse actions on marks.}
\usage{
handle_click(vis, on_click = NULL)

handle_hover(vis, on_mouse_over = NULL, on_mouse_out = NULL)
}
\arguments{
\item{vis}{Visualisation to listen to.}

\item{on_click,on_mouse_over,on_mouse_out}{Callback functions called with
arguments \code{value} and \code{session} every time the mouse is clicked,
mouses over an object, or mouses out.}
}
\description{
Handle mouse actions on marks.
}
\examples{
str_value <- function(value, ...) str(value)
mtcars \%>\% ggvis(~mpg, ~wt) \%>\% layer_points() \%>\%
  handle_click(str_value)
mtcars \%>\% ggvis(~mpg, ~wt) \%>\% layer_points() \%>\%
  handle_hover(str_value, str_value)
}

