% Generated by roxygen2 (4.0.0): do not edit by hand
\name{handler}
\alias{handler}
\alias{is.handler}
\title{Create handler S3 class.}
\usage{
handler(subclass, listener, control_args = list(), value = NULL,
  map = identity, id = rand_id())

is.handler(x)
}
\arguments{
\item{subclass}{name of the subclass. \code{handler} is an abstract base
class so this must always be provided.}

\item{listener}{name of the js listener (with corresponding R event broker)
that this handler is associated with}
}
\description{
This is currently a subclass on input, but it should probably be the
other way around since inputs are handlers that have controls.
}
\examples{
p <- qvis(mtcars, ~mpg, ~wt, size := left_right(1, 100))
p$props$size.update$dr

# Handlers are extracted with the internal handlers() function
# ggvis:::handlers(p)
}

