% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_utils.R
\name{insert}
\alias{insert}
\title{insert a deep value within a list of lists}
\usage{
insert(l, obj, value = NULL)
}
\arguments{
\item{l}{list to update}

\item{obj}{list with the object names to identifying the path to the object to change}

\item{value}{The value to store in the object to change}
}
\value{
list l with the value updated
}
\description{
This can be useful for updating specific settings within the vega object
}
\examples{
test <- list(a=1, b=2, c=list(e=list(g=4, h=5), f=3))
insert(test, list("c", "e", "g"), 7)
insert(test, list("c", "e", 2), 8)
}
