% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.axis_props}
\alias{is.axis_props}
\title{Tests whether an object is an axis_props object}
\usage{
is.axis_props(x)
}
\arguments{
\item{x}{an object to test}
}
\description{
Tests whether an object is an axis_props object
}
\keyword{internal}

