% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.legend_props}
\alias{is.legend_props}
\title{Tests whether an object is a legend_props object}
\usage{
is.legend_props(x)
}
\arguments{
\item{x}{an object to test}
}
\description{
Tests whether an object is a legend_props object
}
\keyword{internal}

