% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_props.R
\name{is.scaled_value}
\alias{is.scaled_value}
\title{Tests whether an object is a scaled_value object}
\usage{
is.scaled_value(x)
}
\arguments{
\item{x}{an object to test}
}
\description{
Tests whether an object is a scaled_value object
}
\keyword{internal}

