% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_is.R
\name{is.vega}
\alias{is.vega}
\title{Validae that the vis includes a valid vega object}
\usage{
is.vega(vis, error = FALSE)
}
\arguments{
\item{vis}{a ggvis vis}

\item{error}{Throw an error on parse failure? If TRUE, then the function returns NULL
on success (i.e., call only for the side-effect of an error on failure, like
stopifnot).}
}
\value{
logical
}
\description{
Validae that the vis includes a valid vega object
}
