% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_is.R
\name{is.vega_config}
\alias{is.vega_config}
\title{Validate if the object is a vega config object}
\usage{
is.vega_config(obj)
}
\arguments{
\item{obj}{an object}
}
\value{
logical
}
\description{
Validate if the object is a vega config object
}
