% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_is.R
\name{is.vega_title}
\alias{is.vega_title}
\title{Validate if the object is a vega title object}
\usage{
is.vega_title(obj)
}
\arguments{
\item{obj}{an object}
}
\value{
logical
}
\description{
Validate if the object is a vega title object
}
