% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.ggvis}
\alias{knit_print.ggvis}
\title{Knit print method for ggvis plots.}
\usage{
knit_print.ggvis(x, options = list(), inline = FALSE, ...)
}
\description{
Knit print method for ggvis plots.
}
\keyword{internal}

