% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/layer_axial.R
\name{layer_axial_lines}
\alias{layer_axial_lines}
\title{Draw vertical or horizontal lines}
\usage{
layer_axial_lines(vis, ..., data = NULL)
}
\arguments{
\item{vis}{Visualisation to modify}

\item{...}{A \code{\link{props}} object, named according to the
properties listed below.}

\item{data}{An optional dataset, if you want to override the usual data
inheritance for this mark.}
}
\description{
This adds axial lines to the plot, that is, lines that are parallel
to the axes. For each \code{x}, a vertical line spanning the
plot is created. Each \code{y} will result in a vertical line.
}
\examples{
# In this example we will represent graphically the means of the
# mixture components of the faithful dataset.
# First we compute the means of each component:
means <- faithful \%>\%
  mutate(left = eruptions < 3) \%>\%
  group_by(left) \%>\%
  summarise(values = mean(eruptions))

# Now we draw vertical lines corresponding to those means, plus an
# horizontal line 100 pixel from the top of the plot
ggvis(faithful) \%>\%
  layer_histograms(~eruptions) \%>\%
  layer_axial_lines(data = means,
    x = ~values, y := 100,
    stroke := "red", strokeWidth := 4
  )
}

