% Generated by roxygen2 (4.0.1): do not edit by hand
\name{layer_boxplot}
\alias{layer_boxplot}
\title{Display data with a boxplot.}
\usage{
layer_boxplot(vis, ..., width = NULL)
}
\arguments{
\item{vis}{Visualisation to modify}

\item{...}{Visual properties used to override defaults.}

\item{width}{Width of each bar. When x is continuous, this controls the width
in the same units as x. When x is categorical, this controls the width as a
proportion of the spacing between items (default is 0.9).}
}
\description{
This will add boxplots to a plot. The action of \code{layer_boxplot} depends
on whether the \code{x} prop is continuous or categorical.
}
\examples{
library(dplyr)

# Make data set with categorical x
mtc <- mtcars
mtc$cyl <- factor(mtc$cyl)

mtc \%>\% ggvis(~cyl, ~mpg) \%>\% layer_boxplot()
# Set the width of the boxes to half the space between tick marks
mtc \%>\% ggvis(~cyl, ~mpg) \%>\% layer_boxplot(width = 0.5)

# Continuous x: boxes fill width between data values
mtcars \%>\% ggvis(~cyl, ~mpg) \%>\% layer_boxplot()
# Setting width=0.5 makes it 0.5 wide in the data space, which is 1/4 of the
# distance between data values in this particular case.
mtcars \%>\% ggvis(~cyl, ~mpg) \%>\% layer_boxplot(width = 0.5)
}
\seealso{
\code{\link{compute_boxplot}} for more information on how data is
  transformed.
}

