% Generated by roxygen2 (4.0.0): do not edit by hand
\name{layer_f}
\alias{layer_f}
\title{Create a new layering function.}
\usage{
layer_f(vis, fun)
}
\arguments{
\item{vis}{The ggvis visualisation to modify.}

\item{fun}{A function that takes a single argument, the current
visualisation as input, and returns a modified visualisation.}
}
\description{
Create a new layering function.
}
\keyword{internal}

