% Generated by roxygen2 (4.0.0): do not edit by hand
\name{layer_guess}
\alias{layer_guess}
\title{Guess the right type of layer based on current properties.}
\usage{
layer_guess(vis, ...)
}
\arguments{
\item{vis}{The visualisation to add the new layer to.}
}
\description{
\code{layer_guess} provides the magic behind the default behaviour of
\code{\link{qvis}}.
}
\section{Defaults}{


\itemize{
  \item Continuous x, \code{\link{layer_histogram}}
  \item Categorical x, \code{\link{layer_barchart}}
  \item Continuous x and y, \code{\link{layer_points}}
}
}
\examples{
# A scatterplot:
mtcars \%>\% ggvis(~mpg, ~wt) \%>\% layer_guess()
# A histogram:
mtcars \%>\% ggvis(~mpg) \%>\% layer_guess()
}

