% Generated by roxygen2 (4.0.0): do not edit by hand
\name{layer_histograms}
\alias{layer_bars}
\alias{layer_freqpolys}
\alias{layer_histograms}
\title{Display binned data}
\usage{
layer_histograms(vis, ..., binwidth = NULL, origin = NULL, right = TRUE)

layer_freqpolys(vis, ..., binwidth = NULL, origin = NULL, right = TRUE)

layer_bars(vis, ...)
}
\arguments{
\item{vis}{Visualisation to modify}

\item{binwidth}{The width of the bins. The default is \code{NULL}, which
yields 30 bins that cover the range of the data. You should always override
this value, exploring multiple widths to find the best to illustrate the
stories in your data.}

\item{origin}{The initial position of the left-most bin. If \code{NULL}, the
the default, will use the smallest value in the dataset.}

\item{right}{Should bins be right-open, left-closed, or
right-closed, left-open}
}
\description{
Display binned data
}
\examples{
# Create histograms and frequency polygons with layers
mtcars \%>\% ggvis(~mpg) \%>\% layer_histograms()
mtcars \%>\% ggvis(~mpg) \%>\% layer_histograms(binwidth = 2)
mtcars \%>\% ggvis(~mpg) \%>\% layer_freqpolys(binwidth = 2)

# These are equivalent to combining compute_bin with the corresponding
# mark
mtcars \%>\% compute_bin("mpg") \%>\% ggvis(~x_, ~count_) \%>\% layer_paths()
}

