% Generated by roxygen2 (4.0.0): do not edit by hand
\name{layer_line}
\alias{layer_line}
\title{A layer to add a line.}
\usage{
layer_line(..., sort = TRUE)
}
\arguments{
  \item{sort}{Should the data be sorted? If \code{TRUE}
  (the default), a \code{\link{transform_sort}} will be
  used. By default, this will sort the data on the x
  variable.}

  \item{...}{Named arguments are passed on to
  \code{\link{transform_sort}}; unnamed arguments are not
  used.}
}
\description{
A layer to add a line.
}
\examples{
ggvis(mtcars,
  props(x = ~wt, y = ~mpg, stroke = ~factor(cyl)),
  layer_line(),
  layer_point()
)
}
\seealso{
\code{link{mark_path}}
}

