% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_line2.R
\name{layer_line_}
\alias{layer_line_}
\title{Generate a line chart}
\usage{
layer_line_(vis, ...)
}
\arguments{
\item{vis}{Visualisation to modify}

\item{...}{Arguments passed on to \code{layer_path_}
\describe{
  \item{vis}{Visualisation to modify}
  \item{from}{The data or other source to use for this figure. May be a dataframe or a name.}
  \item{stack}{If there are multiple bars to be drawn at an x location, should
the bars be stacked? If FALSE, the bars will be overplotted on each other.}
  \item{id}{A string to append to names created within this layer, (for example
data or transform fields)}
}}
}
\description{
Assumptions:
  - y is numeric
  - all columns beyond x, y, and grouping will be aggregated
}
\details{
The only difference between a line and path chart is the line chart is ordered
}
