% Generated by roxygen2 (4.0.0): do not edit by hand
\name{layer_lines}
\alias{layer_lines}
\title{Layer lines on a plot.}
\usage{
layer_lines(vis, ..., sort = TRUE)
}
\arguments{
\item{sort}{Should the data be sorted? If \code{TRUE} (the default), a
\code{\link{transform_sort}} will be used. By default, this will sort the
data on the x variable.}

\item{vis}{Visualisation to modify.}

\item{...}{Named arguments are passed on to \code{\link{transform_sort}};
unnamed arguments are not used.}
}
\description{
Layer lines on a plot.
}
\examples{
mtcars2 <- dplyr::mutate(mtcars, cyl = factor(cyl))
mtcars2 \%>\% ggvis(~wt, ~mpg, stroke = ~cyl) \%>\% layer_lines()

# Equivalent to
mtcars2 \%>\% ggvis(~wt, ~mpg, stroke = ~cyl) \%>\%
  group_by(cyl) \%>\% transform_sort() \%>\% layer_paths()
}
\seealso{
\code{link{mark_path}}
}

