% Generated by roxygen2 (4.0.1): do not edit by hand
\name{legend_props}
\alias{legend_props}
\title{Create an axis_props object for controlling legend properties.}
\usage{
legend_props(title = NULL, labels = NULL, symbols = NULL,
  gradient = NULL, legend = NULL)
}
\arguments{
\item{title}{A named list of text properties for the legend title.}

\item{labels}{A named list of text properties for legend labels.}

\item{symbols}{A named list of line properties for symbols (for discrete
legend items).}

\item{gradient}{A named list of line properties a continuous color gradient.}

\item{legend}{A named list of line properties for the overall legend. The
  x and y position can be set here, which will override automatic
  positioning.}
}
\description{
Create an axis_props object for controlling legend properties.
}

