% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mainTopPanel}
\alias{mainTopPanel}
\title{Create a main panel which moves to the top}
\usage{
mainTopPanel(...)
}
\arguments{
  \item{...}{UI elements to include in the main panel.}
}
\description{
This is to be used with \code{link{sidebarBottomPage}}.
}

