% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prop.R
\name{new_prop}
\alias{new_prop}
\title{Create new prop object}
\usage{
new_prop(x, property, scale, offset, mult, env, event, label)
}
\description{
The resulting object has the following fields:
}
\details{
\itemize{
  \item property The name of a visual property, like "x", "x2", "width", "y",
    "fill".
  \item value A value. Can be a constant, reactive, or quoted expression.
  \item scale A string with name of a scale. Typically something like "x",
    "y", "fill", but can also be a custom name like "foo".
  \item offset Additive pixel offset used to adjust scaled values.
  \item mult Multiplicative pixel offset used to adjust scaled values.
  \item event A event like "update", "enter", "exit", "hover", "brush".
  \item env An environment in which to evaluate a variable or reactive value.
}
}
\keyword{internal}

