% Generated by roxygen2 (4.0.0): do not edit by hand
\name{\%>\%}
\alias{\%>\%}
\alias{as.pipe}
\alias{is.pipe}
\alias{pipe}
\title{Pipe graphics}
\usage{
lhs \%>\% rhs

pipe(type, ...)

is.pipe(x)

as.pipe(x, ...)
}
\arguments{
\item{lhs,rhs}{A visualisation and a function to apply to it}
}
\description{
Like dplyr, ggvis also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.

A pipe object represents a component in a \code{\link{pipeline}}. Pipes
provide a declarative specification of interactive behaviour, and define
the behaviour of each component in the data hierarchy.
}
\details{
This function is designed to be used by authors of new types of pipes.
If you are a ggvis user, please use an existing pipe: a data frame,
a transform, a mark, or a layer
}
\examples{
# Instead of
layer_points(ggvis(mtcars, ~mpg, ~wt))
# you can write
mtcars \%>\% ggvis(~mpg, ~wt) \%>\% layer_points()
}
\keyword{internal}

