% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_.R
\name{print.ggvega}
\alias{print.ggvega}
\title{Render using plot using htmltools and vega standard process}
\usage{
\method{print}{ggvega}(x, launch = interactive(), ...)
}
\arguments{
\item{x}{Visualisation to modify.}

\item{launch}{If \code{TRUE}, will launch plot in a viewer/browser. If
\code{FALSE} returns an object that you can \code{print()} to launch.}

\item{...}{'render', 'logLevel', and 'tooltop_opts' parameters for vega_render() function}
}
\value{
printable object
}
\description{
Render using plot using htmltools and vega standard process
}
\examples{
p <- mtcars \%>\%
  dplyr::group_by(cyl) \%>\%
  dplyr::summarize(mpg = median(mpg)) \%>\%
  dplyr::mutate(cyl = as.factor(cyl)) \%>\%
  ggvis::ggvis(x=~cyl, y=~mpg) \%>\%
  ggvis::layer_bars(fill:="steelblue") \%>\%
  ggvis::add_title(text="MPG by Cylinder", orient="top", anchor="start") \%>\%
  ggvis::flip()
tooltips_opts <- list(
  showAllFields=FALSE,
  fields=list(
    list(
      field= "x_",
      title= "Cylinder",
      formatType= "string"
    ),
    list(
      field= "stack_upr_",
      title= "MPG",
      formatType= "number"
    )
  )
)
vega_render(dump_spec(p), tooltip_opts = tooltips_opts)
}
