% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.ggvis}
\alias{knit_print.ggvis}
\alias{print.ggvis}
\alias{view_dynamic}
\alias{view_static}
\title{View in a ggvis plot in the browser.}
\usage{
\method{print}{ggvis}(x, dynamic = NA, id = rand_id("plot_"),
  minified = TRUE, ...)

view_static(x, plot_id = rand_id("plot_"), minified = TRUE,
  dest = tempfile(pattern = "ggvis"))

view_dynamic(x, plot_id = rand_id("plot_"), minified = TRUE, port = NULL,
  quiet = TRUE)

knit_print.ggvis(x, options = list())
}
\arguments{
\item{x}{A ggvis object.}

\item{dynamic}{Uses \code{view_dynamic} if \code{TRUE}, \code{view_static} if
\code{FALSE}. The default, \code{NA}, chooses automatically based on the
presence of reactives or interactive inputs in \code{x}.}

\item{...}{Other arguments passed on to \code{view_dynamic} and
\code{view_static} from \code{print}.}

\item{launch}{If \code{TRUE}, launch this web page in a browser or Rstudio.}

\item{port}{the port on which to start the shiny app. If NULL (the default),
Shiny will select a random port.}

\item{quiet}{If \code{TRUE} show status messages from Shiny. (Default is
\code{FALSE}.)}

\item{minified}{If \code{TRUE}, use minified version of JS and CSS files. This
can be useful for debugging.}
}
\description{
\code{view_static} creates a static web page in a temporary directory;
\code{view_dynamic} generate a dynamic shiny app and launches it.
\code{print} picks between the two methods automatically based on whether
or not your plot has interactive features.
}
\keyword{internal}

