\name{prop_type}
\alias{prop_type}
\title{Determine the variable type given a data frame and property.}
\usage{
  prop_type(data, prop, processed = FALSE)
}
\arguments{
  \item{data}{The data object.}

  \item{prop}{The property object.}

  \item{processed}{Has this data object been processed so
  that new columns have been calculated and unused columns
  have been dropped?}
}
\description{
  Determine the variable type given a data frame and
  property.
}
\keyword{internal}

