\name{prop_value}
\alias{prop_value}
\title{Given a property and a dataset, get the value of the property.}
\usage{
  prop_value(x, data)
}
\arguments{
  \item{x}{a vega property}

  \item{data}{a object containing data}
}
\description{
  Given a property and a dataset, get the value of the
  property.
}

