\name{prop_vega}
\alias{prop_vega}
\title{Generate a vega object for the individual mark.}
\usage{
  prop_vega(x, default_scale)
}
\arguments{
  \item{x}{a vega property}

  \item{default_scale}{default scale to use if not already
  specified in vega object}
}
\description{
  Will be inserted in to the vega properties object.
}
\keyword{internal}

