% Generated by roxygen2 (4.0.2): do not edit by hand
\name{propname_to_scale}
\alias{propname_to_scale}
\title{Convert the name of a property to the name of its default scale.}
\usage{
propname_to_scale(prop)
}
\arguments{
\item{prop}{character vector of property names. Any unrecognised names
are left unchanged.}
}
\value{
character vector of default scale names.
}
\description{
This is mainly used to ensure that similar properties share the same
scale by default - e.g. \code{x} and \code{x2} should use the same
scale.
}
\examples{
propname_to_scale(c("x", "x2"))
propname_to_scale(c("foo", "bar"))
propname_to_scale(c("opacity", "fillOpacity", "strokeOpacity"))
}
\keyword{internal}

