\name{proptype_to_scale}
\alias{proptype_to_scale}
\title{Convert the type of a property to the type of it's default scale.}
\usage{
  proptype_to_scale(type)
}
\arguments{
  \item{type}{property type: numeric, ordinal, nominal,
  logical or datetime.}
}
\description{
  Convert the type of a property to the type of it's
  default scale.
}
\keyword{internal}

