% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{Resize-class}
\alias{Resize}
\alias{Resize-class}
\title{Event broker for resize events.}
\description{
The resize event broker is useful if you want your shiny app to respond
to resize events.
}
\section{Methods}{

\describe{
\item{\code{resize()}}{A reactive value changes when the plot is resized.
Returns a list containing:
  * plot_id: The ID of the ggvis plot.
  * width: Width of the plot in pixels.
  * height: Height of the plot in pixels.
  * padding: A list containing `left`, `right`, `top`, `bottom`, which is
    the padding in pixels for each side. The padding is inside of the
    width and height values.
  }
}}

