% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scale_datetime}
\alias{scale_datetime}
\title{Add a date-time scale to a ggvis object.}
\usage{
scale_datetime(vis, scale, domain = NULL, range = NULL, reverse = FALSE,
  round = FALSE, utc = FALSE, clamp = FALSE, nice = NULL, name = NULL)
}
\arguments{
\item{vis}{A ggvis object.}

\item{scale}{The name of a scale, such as "x", "y", "fill", "stroke", etc.}

\item{clamp}{If true, values that exceed the data domain are clamped to
either the minimum or maximum range value.}

\item{nice}{If specified, modifies the scale domain to use a more
human-friendly value range. Should be a string indicating the desired time
interval; legal values are "second", "minute", "hour", "day", "week",
"month", or "year"}

\item{utc}{if \code{TRUE}, uses UTC times.}

\item{domain}{The domain of the scale, representing the set of data values.
For ordinal scales, a character vector; for quantitative scales, a numeric
vector of length two. Either value (but not both) may be NA, in which
case \code{domainMin} or \code{domainMax} is set.}

\item{range}{The range of the scale, representing the set of visual values.
For numeric values, the range can take the form of a two-element array with
minimum and maximum values. For ordinal data, the range may by an array of
desired output values, which are mapped to elements in the specified
domain. The following range literals are also available: "width", "height",
"shapes", "category10", "category20".}

\item{reverse}{If true, flips the scale range.}

\item{round}{If true, rounds numeric output values to integers. This can be
helpful for snapping to the pixel grid.}

\item{name}{name of the scale.}
}
\description{
A date/time scale controls the mapping of date and time variables to
visual properties.
}
\examples{
set.seed(2934)
dat <- data.frame(
  time = as.Date("2013-07-01") + 1:100,
  value = seq(1, 10, length.out = 100) + rnorm(100)
)
p <- dat \%>\% ggvis(~time, ~value) \%>\% layer_points()

# Start and end on month boundaries
p \%>\% scale_datetime("x", nice = "month")


dist <- data.frame(times = as.POSIXct("2013-07-01", tz = "GMT") +
                           rnorm(200) * 60 * 60 * 24 * 7)
p <- dist \%>\% ggvis(x = ~times) \%>\% layer_histograms()
p

# Start and end on month boundaries
p \%>\% scale_datetime("x", nice = "month")

p \%>\% scale_datetime("x", utc = TRUE)
}
\seealso{
\code{\link{scales}}, \code{\link{scale_numeric}},
  \url{https://github.com/trifacta/vega/wiki/Scales#time-scale-properties}

Other scales: \code{\link{scale_logical}},
  \code{\link{scale_nominal}}, \code{\link{scale_ordinal}};
  \code{\link{scale_numeric}}
}

