% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scale_info}
\alias{scale_info}
\title{Create a scale_info object. These objects are typically associated with a
mark, and they contain information about the scale. Note that these are
different from vega scale objects.}
\usage{
scale_info(scale, label, type, domain, override = FALSE)
}
\arguments{
\item{label}{Label for the scale.}

\item{type}{Type of scale.}

\item{domain}{Can be either a vector of values for the domain, or a reactive
which returns such a vector.}

\item{override}{Should the domain specified by this scale_info object
override other scale_infos for the same scale? Useful when domain is
manually specified.}
}
\description{
Create a scale_info object. These objects are typically associated with a
mark, and they contain information about the scale. Note that these are
different from vega scale objects.
}
\keyword{internal}

