% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_props.R
\name{scaled_value}
\alias{scaled_value}
\title{Create a scaled_value object}
\usage{
scaled_value(scale, value)
}
\arguments{
\item{scale}{The name of a scale, e.g., "x", "fill".}

\item{value}{A value which will be transformed using the scale.}
}
\description{
These are for use with legends and axes.
}
