\name{scales}
\alias{scales}
\title{Create a gigvis_scales object}
\usage{
  scales(..., .scales = list())
}
\arguments{
  \item{...,.scales}{scales to combine into a single scales
  object}
}
\description{
  Given arguments which are scales object, return a list
  with those objects, where the entries also have names
  that are the same as the "name" field of the objects.
}
\keyword{internal}

