% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scaletype_to_vega_scaletype}
\alias{scaletype_to_vega_scaletype}
\title{Given the type of a ggvis scale, get the name of its corresponding vega scale}
\usage{
scaletype_to_vega_scaletype(type)
}
\arguments{
\item{type}{property type: numeric, ordinal, nominal, logical or datetime.}
}
\description{
Given the type of a ggvis scale, get the name of its corresponding vega scale
}
\keyword{internal}

