% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{set_scale_label}
\alias{set_scale_label}
\title{Set the label for a scale}
\usage{
set_scale_label(vis, scale, label)
}
\arguments{
\item{vis}{A ggvis object.}

\item{scale}{The name of a scale, like "x".}

\item{label}{Text to use for the label.}
}
\description{
Set the label for a scale
}

