% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny_layout.R
\name{sidebarBottomPage}
\alias{mainTopPanel}
\alias{sidebarBottomPage}
\alias{sidebarBottomPanel}
\title{Create a page with a sidebar}
\usage{
sidebarBottomPage(sidebarPanel, mainPanel, shiny_headers = TRUE)

sidebarBottomPanel(...)

mainTopPanel(...)
}
\arguments{
\item{sidebarPanel}{The \code{\link{sidebarBottomPanel}} containing input
controls.}

\item{mainPanel}{The \code{\link{mainTopPanel}} containing the main content.}

\item{shiny_headers}{Should Shiny headers be embedded in the page? This
should be TRUE for interactive/dynamic pages, FALSE for static pages.}

\item{...}{Additional tags.}
}
\description{
This creates a page with a sidebar, where the sidebar moves to the bottom
when the width goes below a particular value.
}
\examples{
sidebarBottomPage(sidebarBottomPanel(), mainTopPanel())
}

