% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sidebarBottomPanel}
\alias{sidebarBottomPanel}
\title{Create a sidebar panel which moves to the bottom}
\usage{
sidebarBottomPanel(...)
}
\arguments{
  \item{...}{UI elements to include in the sidebar.}
}
\description{
This is to be used with \code{link{sidebarBottomPage}}.
}

