% Generated by roxygen2 (4.0.1): do not edit by hand
\name{singleton}
\alias{scale_singleton}
\alias{singleton}
\title{Singleton.}
\usage{
singleton()

scale_singleton(vis, scale, points = TRUE)
}
\arguments{
\item{vis}{A ggvis object.}

\item{scale}{The name of a scale, such as "x", "y", "fill", "stroke", etc.}

\item{points}{If \code{TRUE}, distributes the ordinal values over a
quantitative range at uniformly spaced points. The spacing of the points
can be adjusted using the padding property. If \code{FALSE}, the ordinal
scale will construct evenly-spaced bands, rather than points.}
}
\description{
Use singleton when you want constant x or y position.
}
\examples{
mtcars \%>\% ggvis("", ~mpg) \%>\%
  layer_points() \%>\%
  scale_nominal("x") \%>\%
  add_guide_axis("x", title = "", tick_size_major = 0)

# OR
mtcars \%>\% ggvis("", ~mpg) \%>\%
  layer_points() \%>\%
  scale_singleton("x")

# OR, even simpler
mtcars \%>\% ggvis(singleton(), ~mpg) \%>\% layer_points()

# In the other direction:
mtcars \%>\% ggvis(~mpg, singleton()) \%>\% layer_points()
}

