% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subvis}
\alias{subvis}
\title{Create a subvisualisation.}
\usage{
subvis(vis, ..., layer)
}
\description{
A subvis is a recursive mark: a mark that contain other marks. Compared
to other marks, it can also contain scales, axes and legends.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

small <- nasaweather::atmos \%>\%
  filter(lat <= -11.217391, long <= -106.287, year == 1995) \%>\%
  group_by(long, lat)
small \%>\%
  ggvis(~long, ~lat) \%>\%
  subvis(width := 20, height := 20, stroke := "red",
    layer = function(x) x \%>\% layer_points(~month, ~ozone))
}

