% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_.R
\name{update_props}
\alias{update_props}
\title{update properties from ggvega, a signal, transform, or data source}
\usage{
update_props(properties, from = NA, ...)
}
\arguments{
\item{properties}{properties data frame (from vis$properties)}

\item{from}{string the name of the source if applicable. e.g. 'mtcars'}

\item{...}{parameters (e.g. x=cyl, y=mpg, fill="blue")}
}
\value{
a modified visualization
}
\description{
update properties from ggvega, a signal, transform, or data source
}
