% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ref.R
\name{update_signal_}
\alias{update_signal_}
\title{Update a vega signal in a vega vis}
\usage{
update_signal_(vis, ...)
}
\arguments{
\item{vis}{a ggvis object}

\item{...}{vega signal properties}
}
\value{
a ggvis object
}
\description{
if the signal name isn't found int he vis, it is added
}
