% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_bin.R
\name{vega_bin_transform}
\alias{vega_bin_transform}
\title{create a vega bin transform object}
\usage{
vega_bin_transform(type, field = NULL, extent = NULL, anchor = NULL,
  maxbins = NULL, base = NULL, step = NULL, steps = NULL,
  minstep = NULL, divide = NULL, nice = NULL, signal = NULL,
  as = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{field}{Field (required) The data field to bin.}

\item{extent}{Number[] (required) A two-element array with the minimum and maximum values of the bin range.}

\item{anchor}{Number A value in the binned domain at which to anchor the bins, shifting the bin boundaries if necessary to ensure that a boundary aligns with the anchor value. By default, the minimum bin extent value serves as the anchor.}

\item{maxbins}{Number The maximum number of bins to create (default `20`).}

\item{base}{Number The number base to use for automatic bin determination (default `10`).}

\item{step}{Number An exact step size to use between bins. If provided, options such as _maxbins_ will be ignored.}

\item{steps}{Number[] An array of allowable step sizes to choose from.}

\item{minstep}{Number The minimim allowed bin step size (default `0`).}

\item{divide}{Number[] Allowable bin step sub-divisions. The default value is `[5, 2]`, which indicates that for base 10 numbers (the default base) automatic bin determination can consider dividing bin step sizes by 5 and/or 2.}

\item{nice}{Boolean If `true` (the default), attempts to make the bin boundaries use human-friendly boundaries, such as multiples of ten.}

\item{signal}{String If defined, binds the computed binning specification (an object with _start_, _stop_ and _step_ properties) to a signal with the given name.}

\item{as}{String[] The output fields at which to write the start and end bin values. The default is `["bin0", "bin1"]`.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/bin/index.html
}
\details{
The **bin** transform discretizes numeric values into a set of bins. A common use case is to create a histogram.
}
