% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_collect.R
\name{vega_collect_transform}
\alias{vega_collect_transform}
\title{create a vega collect transform object}
\usage{
vega_collect_transform(type, sort = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{sort}{Compare A comparator definition for sorting data objects.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/collect/index.html
}
\details{
The **collect** transform collects all the objects in a data stream within a single array, allowing sorting by data field values.
}
