% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_contour.R
\name{vega_contour_transform}
\alias{vega_contour_transform}
\title{create a vega contour transform object}
\usage{
vega_contour_transform(type, size = NULL, values = NULL, x = NULL,
  y = NULL, cellSize = NULL, bandwidth = NULL, thresholds = NULL,
  count = NULL, nice = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{size}{Number[] (required) The dimensions `[width, height]` over which to compute contours. If the values parameter is provided, this must be the dimensions of the input data. If density estimation is performed, this is the output view dimensions in pixels.}

\item{values}{Number[]  An array of numeric values representing an width x height grid of values over which to compute contours. If unspecified, this transform will instead compute contours for the kernel density estimate of input data.}

\item{x}{Field The pixel x-coordinate field for density estimation.}

\item{y}{Field The pixel y-coordinate field for density estimation.}

\item{cellSize}{Number Contour density calculation cell size.}

\item{bandwidth}{Number Kernel density estimation bandwidth.}

\item{thresholds}{Number[] A threshold array of explicit contour boundaries. If this parameter is set, the _count_ and _nice_ parameters will be ignored.}

\item{count}{Number The desired number of contours. This parameter is ignored if the _thresholds_ parameter is provided.}

\item{nice}{Boolean A boolean flag (default `false`) indicating if the contour threshold values should be automatically aligned to "nice", human-friendly values. Setting this flag may cause the number of thresholds to deviate from the exact _count_.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/contour/index.html
}
\details{
The **contour** transform models a spatial distribution of data values using a set of discrete levels. Each [contour line](https://en.wikipedia.org/wiki/Contour_line) is an isoline of constant value. A common use case is to convey density estimates for 2D point data, as these can provide a more scalable representation for large numbers of data points.

The contour transform generates a new stream of [GeoJSON](https://en.wikipedia.org/wiki/GeoJSON) geometry data as output. These shapes can then be visualized using either the [geoshape](../geoshape) or [geopath](../geopath) transform. For a complete example, see the [contour plot example visualization](../../../examples/contour-plot). This transform provides the functionality of both the [contours](https://github.com/d3/d3-contour/#contours) and [densityContour](https://github.com/d3/d3-contour/#densityContour) methods of the [d3-contour](https://github.com/d3/d3-contour) module.
}
