% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_cross.R
\name{vega_cross_transform}
\alias{vega_cross_transform}
\title{create a vega cross transform object}
\usage{
vega_cross_transform(type, filter = NULL, as = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{filter}{Expr An optional filter expression for limiting the results of the cross-product.}

\item{as}{Array The output fields for the two data objects being crossed. The default is `["a", "b"]`.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/cross/index.html
}
\details{
The **cross** transform compute the cross-product of a data stream with itself.
}
