% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_type.R
\name{vega_data_parser}
\alias{vega_data_parser}
\title{Determine the vega data type for a vector}
\usage{
vega_data_parser(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
This is used to specify the data type so that the appropriate parser is used
when Vega receives the data.
}

