% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_extent.R
\name{vega_extent_transform}
\alias{vega_extent_transform}
\title{create a vega extent transform object}
\usage{
vega_extent_transform(type, field = NULL, signal = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{field}{Field (required) The data field for which to compute the extent.}

\item{signal}{String If defined, binds the computed extent array to a signal with the given name.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/extent/index.html
NOTE: Some parameters are required to be 'arrays'.  In the case where a parameter must be an array but is a single line, wrap it in 'I()', (asis), to ensure it is properly handled.
}
\details{
The **extent** transform computes the minimum and maximum values for a data field, producing a `[min, max]` array. This transform is useful for computing a value range and binding it to a signal name, for example to use as a parameter for a [bin](../bin) transform. This transform does not change the input data stream, it only computes the extent as a side effect.
}
