% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_fold.R
\name{vega_fold_transform}
\alias{vega_fold_transform}
\title{create a vega fold transform object}
\usage{
vega_fold_transform(type, fields = NULL, as = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{fields}{Field[] (required) An array of data fields indicating the properties to fold.}

\item{as}{String[] The output field names for the _key_ and _value_ properties produced by the fold transform. The default is `["key", "value"]`.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/fold/index.html
NOTE: Some parameters are required to be 'arrays'.  In the case where a parameter must be an array but is a single line, wrap it in 'I()', (asis), to ensure it is properly handled.
}
\details{
The **fold** transform collapses (or "folds") one or more data fields into two properties: a _key_ property (containing the original data field name) and a _value_ property (containing the data value). The fold transform is useful for mapping matrix or cross-tabulation data into a standardized format.

This transform generates a new data stream in which each data object consists of the _key_ and _value_ properties as well as all the original fields of the corresponding input data object.
}
