% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_formula.R
\name{vega_formula_transform}
\alias{vega_formula_transform}
\title{create a vega formula transform object}
\usage{
vega_formula_transform(type, expr = NULL, as = NULL, initonly = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{expr}{Expr (required) The formula [expression](../../expressions) for calculating derived values.}

\item{as}{String (required) The output field at which to write the formula value.}

\item{initonly}{Boolean If `true`, the formula is evaluated only when a data object is first observed. The formula values will _not_ automatically update if data objects are modified. Th default is `false`.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/formula/index.html
NOTE: Some parameters are required to be 'arrays'.  In the case where a parameter must be an array but is a single line, wrap it in 'I()', (asis), to ensure it is properly handled.
}
\details{
The **formula** transform extends data objects with new values according to a calculation formula.
}
