% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_geojson.R
\name{vega_geojson_transform}
\alias{vega_geojson_transform}
\title{create a vega geojson transform object}
\usage{
vega_geojson_transform(type, fields = NULL, geojson = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{fields}{Field[] Data fields containing  longitude and latitude values, respectively.}

\item{geojson}{Field Data field containing  GeoJSON feature objects.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/geojson/index.html
}
\details{
The **geojson** transform consolidates geographic data into a single [GeoJSON](https://en.wikipedia.org/wiki/GeoJSON) feature collection. The generated GeoJSON data can then be used to parameterize other parts of a Vega specification, namely the [projection `fit` parameter](../../projections/). This transform can process both latitude / longitude data and existing GeoJSON features.
}
