% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_geopath.R
\name{vega_geopath_transform}
\alias{vega_geopath_transform}
\title{create a vega geopath transform object}
\usage{
vega_geopath_transform(type, projection = NULL, field = NULL, as = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{projection}{String (required) The name of the projection to use.}

\item{field}{Field The data field containing GeoJSON data. If unspecified, the full input data object will be used.}

\item{as}{String The output field to write. The default is `"path"`.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/geopath/index.html
NOTE: Some parameters are required to be 'arrays'.  In the case where a parameter must be an array but is a single line, wrap it in 'I()', (asis), to ensure it is properly handled.
}
\details{
The **geopath** transform maps [GeoJSON](https://en.wikipedia.org/wiki/GeoJSON) features to [SVG path strings](https://developer.mozilla.org/en-US/docs/Web/SVG/Tutorial/Paths) according to a provided cartographic [projection](../../projections). It is intended for use with the [path](../../marks/path) mark type. This transform is similar in functionality to the [geoshape](../geoshape) transform, but immediately generates SVG path strings, rather than producing a shape instance that delays projection until the rendering stage. The [geoshape](../geoshape) transform may have better performance for the case of canvas-rendered dynamic maps.

This transform uses the [d3-geo](https://github.com/d3/d3-geo) library.
}
