% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_geopoint.R
\name{vega_geopoint_transform}
\alias{vega_geopoint_transform}
\title{create a vega geopoint transform object}
\usage{
vega_geopoint_transform(type, projection = NULL, fields = NULL, as = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{projection}{String (required) The name of the projection to use.}

\item{fields}{Field[] (required) The data fields containing the longitude and latitude values, respectively.}

\item{as}{String[] The output fields to write. The default is `["x", "y"]`.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/geopoint/index.html
NOTE: Some parameters are required to be 'arrays'.  In the case where a parameter must be an array but is a single line, wrap it in 'I()', (asis), to ensure it is properly handled.
}
\details{
The **geopoint** transform projects (longitude, latitude) pairs to (x, y) coordinates according to a given cartographic [projection](../../projections).
}
