% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_geoshape.R
\name{vega_geoshape_transform}
\alias{vega_geoshape_transform}
\title{create a vega geoshape transform object}
\usage{
vega_geoshape_transform(type, projection = NULL, field = NULL, as = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{projection}{String (required) The name of the projection to use.}

\item{field}{Field The data field containing GeoJSON data. If unspecified, the full input data object will be used.}

\item{as}{String The output field at which to write the generated shape instance. The default is `"shape"`.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/geoshape/index.html
}
\details{
The **geoshape** transform generates a renderer instance that maps [GeoJSON](https://en.wikipedia.org/wiki/GeoJSON) features to a shape instance that issues drawing commands. It is intended for use solely with the [shape](../../marks/shape) mark type. This transform is similar in functionality to the [geopath](../geopath) transform, but rather than generate intermediate SVG path strings, this transform produces a shape instance that directly generates drawing commands during rendering. This transform can result in improved performance when using canvas rendering for dynamic maps.

This transform uses the [d3-geo](https://github.com/d3/d3-geo) library.
}
