% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_graticule.R
\name{vega_graticule_transform}
\alias{vega_graticule_transform}
\title{create a vega graticule transform object}
\usage{
vega_graticule_transform(type, field = NULL, extentMajor = NULL,
  extentMinor = NULL, extent = NULL, stepMajor = NULL, stepMinor = NULL,
  step = NULL, precision = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{field}{Field The data field to bin.}

\item{extentMajor}{Array[] The major extent of the graticule as a two-element array of coordinates.}

\item{extentMinor}{Array[] The minor extent of the graticule as a two-element array of coordinates.}

\item{extent}{Array[] Sets both the major and minor extents to the same values.}

\item{stepMajor}{Number[] The major step angles of the graticule (default `[90, 360]`).}

\item{stepMinor}{Number[] The minor step angles of the graticule (default `[10, 10]`).}

\item{step}{Number[] Sets both the major and minor step angles to the same values.}

\item{precision}{Number The precision of the graticule in degrees (default `2.5`).}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/graticule/index.html
NOTE: Some parameters are required to be 'arrays'.  In the case where a parameter must be an array but is a single line, wrap it in 'I()', (asis), to ensure it is properly handled.
}
\details{
The **graticule** transform generates a reference grid for cartographic maps. A graticule is a uniform grid of meridians and parallels for showing projection distortion. The default graticule has meridians and parallels every 10° between ±80° latitude; for the polar regions, there are meridians every 90°.

This transform generates a new data stream containing a single [GeoJSON](https://en.wikipedia.org/wiki/GeoJSON) data object for the graticule, which can subsequently be drawn using the [geopath](../geopath) or [geoshape](../geoshape) transform. This transform uses the [d3-geo](https://github.com/d3/d3-geo) library.
}
