% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_identifier.R
\name{vega_identifier_transform}
\alias{vega_identifier_transform}
\title{create a vega identifier transform object}
\usage{
vega_identifier_transform(type, as = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{as}{String (required) The output field at which to write the unique identifier value.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/identifier/index.html
}
\details{
The **identifier** transform extends data objects with a globally unique key value. Identifier values are assigned using an internal counter. This counter is shared across all instances of this transform within a single Vega view, including different data sources. Note, however, that the counter is _not_ shared across different Vega views.
}
