% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_impute.R
\name{vega_impute_transform}
\alias{vega_impute_transform}
\title{create a vega impute transform object}
\usage{
vega_impute_transform(type, field = NULL, key = NULL, keyvals = NULL,
  method = NULL, groupby = NULL, value = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{field}{Field (required) The data field for which missing values should be imputed.}

\item{key}{Field (required) A key field that uniquely identifies data objects within a group. Missing _key_ values (those occuring in the data but not in the current group) will be imputed.}

\item{keyvals}{Any[] An optional array of key values that should be considered for imputation. If provided, this array will be used in addition to the key values observed within the input data.}

\item{method}{String The imputation method to use for the _field_ value of imputed data objects. One of `value` (default), `mean`, `median`, `max` or `min`.}

\item{groupby}{Field[] An optional array of fields by which to group the values. Imputation will then be performed on a per-group basis. For example, missing values may be imputed using the group mean rather than the global mean.}

\item{value}{Any The field value to use when the imputation method is `value`.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/impute/index.html
}
\details{
The **impute** transform performs imputation of missing data objects.
}
