% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_nest.R
\name{vega_nest_transform}
\alias{vega_nest_transform}
\title{create a vega nest transform object}
\usage{
vega_nest_transform(type, generate = NULL, keys = NULL, key = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{generate}{Boolean A boolean flag (default `false`) indicating if non-leaf nodes generated by this transform should be included in the output data stream. By default only the input data objects (leaf nodes) are included. If non-leaf nodes are included, be aware that they will have a different set of properties than the input objects: the generated internal nodes will have a `key` property corresponding to the values extracted using the _keys_ transform parameter.}

\item{keys}{Field[] An array of data fields by which to organize the data objects into a tree. Data objects will first be grouped into siblings based on the values of the first field in the parameter array. The process then continues, creating an additional depth level for each provided field. If unspecified, all data objects will treated as a single set of sibling nodes.}

\item{key}{Field A data field containing a unique key (identifier) for each node. If not specified, Vega will use internal data object ids.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/nest/index.html
}
\details{
The **nest** transform generates a hierarchical (tree) data structure from input data objects, based on dividing children into groups based on distinct field values. Internally, this transform generates a set of tree node objects that can then be processed by tree layout methods such as [tree](../tree), [treemap](../treemap), [pack](../pack), and [partition](../partition).
}
