% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_pie.R
\name{vega_pie_transform}
\alias{vega_pie_transform}
\title{create a vega pie transform object}
\usage{
vega_pie_transform(type, field = NULL, startAngle = NULL, endAngle = NULL,
  sort = NULL, as = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{field}{Field The data values from this field will be encoded as angular spans. If omitted, all pie slices will have equal spans.}

\item{startAngle}{Number The starting angle of the pie in radians (default `0`).}

\item{endAngle}{Number The end angle of the pie in radians (default `2 * PI`).}

\item{sort}{Boolean If true, sorts the arcs according to field values (default `false`).}

\item{as}{String[] The output fields for the computed start and end angles for each arc. The default is `["startAngle", "endAngle"]`.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/pie/index.html
}
\details{
The **pie** transform calculates the angular extents of arc segments laid out in a circle. The most common use case is to create pie charts and donut charts. This transform writes two properties to each datum, indicating the starting and ending angles (in radians) of an arc.
}
