% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{vega_render}
\alias{vega_render}
\title{Render using plot using htmltools and vega standard process}
\usage{
vega_render(vis, renderer = "svg")
}
\arguments{
\item{vis}{ggvis object}

\item{renderer}{"svg" (default) or "canvas"}
}
\value{
printable object
}
\description{
Render using plot using htmltools and vega standard process
}
