% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_resolvefilter.R
\name{vega_resolvefilter_transform}
\alias{vega_resolvefilter_transform}
\title{create a vega resolvefilter transform object}
\usage{
vega_resolvefilter_transform(type, filter = NULL, ignore = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{filter}{Signal (required) A signal reference to a backing [crossfilter](../crossfilter) transform.}

\item{ignore}{Number (required) A bit mask indicating which fields in the crossfilter should be ignored. Each bit corresponds to a field and query in the crossfilter transform's _fields_ and _query_ arrays. If the corresponding bit is on, that field and query will be ignored when resolving the filter. All other queries must pass the filter for a tuple to be included down stream.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/resolvefilter/index.html
}
\details{
The **resolvefilter** transform uses the filter mask generated by a [crossfilter](../crossfilter) transform to efficiently generate filtered data streams. A single crossfilter instance can be used to support cross-filtering over multiple charts; the filtered data streams for each individual chart are generated by using one resolvefilter transform per chart.
}
