% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_sample.R
\name{vega_sample_transform}
\alias{vega_sample_transform}
\title{create a vega sample transform object}
\usage{
vega_sample_transform(type, size = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{size}{Number The maximum number of data objects to include in the sample. The default value is `1000`.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/sample/index.html
}
\details{
The **sample** transform randomly samples a data stream to create a smaller stream. As input data objects are added and removed, the sampled values may change in first-in, first-out manner. This transform uses [reservoir sampling](https://en.wikipedia.org/wiki/Reservoir_sampling) to maintain a representative sample of the stream.
}
