% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_sequence.R
\name{vega_sequence_transform}
\alias{vega_sequence_transform}
\title{create a vega sequence transform object}
\usage{
vega_sequence_transform(type, start = NULL, stop = NULL, step = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{start}{Number (required) The starting value of the sequence.}

\item{stop}{Number (required) The ending value (exclusive) of the sequence.}

\item{step}{Number The step value between sequence entries (default `1`, or `-1` if _stop < start_).}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/sequence/index.html
}
\details{
The **sequence** transform generates a data stream containing a seqence of numeric values. See also the [sequence expression function](../../expressions/#sequence).
}
