% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_signal.R
\name{vega_signal}
\alias{vega_signal}
\title{Add a vega signal}
\usage{
vega_signal(name = NULL, bind = NULL, description = NULL, on = NULL,
  push = NULL, update = NULL, react = TRUE, value = NULL)
}
\arguments{
\item{name}{String Required. A unique name for the signal. Signal names should be valid JavaScript identifiers: they should contain only alphanumeric characters (or “$”, or “_”) and may not start with a digit. Reserved keywords that may not be used as signal names are "datum", "event", and "parent".}

\item{bind}{Bind Binds the signal to an external input element such as a slider, selection list or radio button group.}

\item{description}{String A text description of the signal, useful for inline documentation.}

\item{on}{Handler[] An array of event stream handlers for updating the signal value in response to input events.}

\item{push}{String    Required. To indicate an update to a signal defined in an outer scope, the push property must be set to "outer".}

\item{update}{Expression An update expression for the value of the signal. This expression may include other signals, in which case the signal will automatically update in response to upstream signal changes, so long as the react property is not false.}

\item{react}{Boolean A boolean flag (default true) indicating if the update expression should be automatically re-evaluated when any upstream signal dependencies update. If false, the update expression will only be run upon initialization.}

\item{value}{Any The initial value of the signal (default undefined).}
}
\value{
a vega signal object
}
\description{
https://vega.github.io/vega/docs/signals/
}
\details{
Signals are dynamic variables that parameterize a visualization and can drive interactive behaviors. Signals can be used throughout a Vega specification, for example to define a mark property or data transform parameter.
Signal values are reactive: they can update in response to input event streams, external API calls, or changes to upstream signals. Event streams capture and sequence input events, such as mousedown or touchmove. When an event occurs, signals with associated event handlers are re-evaluated in their specification order. Updated signal values then propagate to the rest of the specification, and the visualization is re-rendered automatically.
}
