\name{vega_spec}
\alias{vega_spec}
\title{Given a filled gigvis spec object, output a gigvis_prerender object}
\usage{
  vega_spec(x, nodes, data_table, width = 600,
    height = 400, padding = NULL, envir = parent.frame())
}
\arguments{
  \item{envir}{The environment in which to evaluate the
  \code{data} parameter of the gigvis object.}
}
\description{
  The gigvis_prerender object contains a Vega spec.
}

