% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_stack.R
\name{vega_stack_transform}
\alias{vega_stack_transform}
\title{create a vega stack transform object}
\usage{
vega_stack_transform(type, field = NULL, groupby = NULL, sort = NULL,
  offset = NULL, as = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{field}{Field The data field that determines the stack heights.}

\item{groupby}{Field[] An array of fields by which to partition the data into separate stacks.}

\item{sort}{Compare Criteria for sorting values within each stack.}

\item{offset}{Number The baseline offset. One of "zero" (default), "center", or "normalize". The "center" offset will center the stacks. The "normalize" offset will compute percentage values for each stack point, with output values in the range [0,1].}

\item{as}{String[] The output fields for the computed start and end stack values. The default is `["y0", "y1"]`.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/stack/index.html
NOTE: Some parameters are required to be 'arrays'.  In the case where a parameter must be an array but is a single line, wrap it in 'I()', (asis), to ensure it is properly handled.
}
\details{
The **stack** transform computes a layout by stacking groups of values. The most common use case is to create stacked graphs, including stacked bar charts and stream graphs. This transform writes two properties to each datum, indicating the starting and ending stack values.
}
