% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_stratify.R
\name{vega_stratify_transform}
\alias{vega_stratify_transform}
\title{create a vega stratify transform object}
\usage{
vega_stratify_transform(type, key = NULL, parentKey = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{key}{Field (required) A data field containing a unique key (identifier) for each node.}

\item{parentKey}{Field (required) A data field containing the key value for each node's parent in the hierarchy.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/stratify/index.html
NOTE: Some parameters are required to be 'arrays'.  In the case where a parameter must be an array but is a single line, wrap it in 'I()', (asis), to ensure it is properly handled.
}
\details{
The **stratify** transform generates a hierarchical (tree) data structure from input data objects, based on key fields that match parent and children nodes. Internally, this transform generates a set of tree node objects that can then be processed by tree layout methods such as [tree](../tree), [treemap](../treemap), [pack](../pack), and [partition](../partition).
}
