% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_treelinks.R
\name{vega_treelinks_transform}
\alias{vega_treelinks_transform}
\title{create a vega treelinks transform object}
\usage{
vega_treelinks_transform(type, key = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{key}{Field A data field containing a unique key (identifier) for each node. This must be the same field used by the upstream [nest](../nest) or [stratify](../stratify) transform.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/treelinks/index.html
NOTE: Some parameters are required to be 'arrays'.  In the case where a parameter must be an array but is a single line, wrap it in 'I()', (asis), to ensure it is properly handled.
}
\details{
The **treelinks** transform generates a new stream of data objects representing links among nodes in a tree. This transform must occur downstream of a tree-generating transform such as [nest](../nest) or [stratify](../stratify). The generated link objects will have `source` and `target` fields that reference input data objects corresponding to parent (source) and child (target) nodes.
}
