% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_wordcloud.R
\name{vega_wordcloud_transform}
\alias{vega_wordcloud_transform}
\title{create a vega wordcloud transform object}
\usage{
vega_wordcloud_transform(type, font = NULL, fontStyle = NULL,
  fontWeight = NULL, fontSize = NULL, fontSizeRange = NULL,
  padding = NULL, rotate = NULL, text = NULL, spiral = NULL,
  as = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{font}{String|Expr The font family to use for a word.}

\item{fontStyle}{String|Expr The font style to use for a word.}

\item{fontWeight}{String|Expr The font weight to use for a word.}

\item{fontSize}{Number|Expr The font size in pixels to use for a word.}

\item{fontSizeRange}{Number[] The range of font sizes to use for the words. If the range is specified and the _fontSize_ is not a numeric constant, the _fontSize_ values will automatically be scaled to lie in the range according to a square root scale.}

\item{padding}{Number|Expr The padding in pixels to place around a word.}

\item{rotate}{Number|Expr The angle in degrees to use for a word.}

\item{text}{Field The data field with the word text.}

\item{spiral}{String The spiral layout method used to place words. One of `archimedean` (the default) or `rectangular`.}

\item{as}{String[] The output fields written by the transform. The default is `["x", "y", "font", "fontSize", "fontStyle", "fontWeight", "angle"]`}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/wordcloud/index.html
}
\details{
The **wordcloud** transform computes a word cloud layout, similar to [Wordle](http://www.wordle.net/). The wordcloud transform is intended for visualizing words or phrases with the [text mark type](../../marks/text). This transform uses [Jason Davies' wordcloud implementation](https://www.jasondavies.com/wordcloud/).
}
