% Generated by roxygen2 (4.0.0): do not edit by hand
\name{add_resize}
\alias{add_resize}
\title{An interactive input bound to resize events.}
\usage{
add_resize(vis, f, id = rand_id())
}
\arguments{
\item{f}{A function which is called each time the plot is resized.}
}
\description{
An interactive input bound to resize events.
}
\examples{
\donttest{
# This example just prints out the current dimensions to the console
print_info <- function(x) {
  str(x)
}

mtcars \%>\% ggvis(~mpg, ~wt) \%>\% layer_points() \%>\% add_resize(print_info)
}
}

