% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggvis.R
\name{add_scale}
\alias{add_scale}
\title{Add arbitrary scales to ggvis.}
\usage{
add_scale(vis, scale, data_domain = TRUE)
}
\arguments{
\item{vis}{Visualisation to modify.}

\item{scale}{Scale object}

\item{data_domain}{Should the domain be controlled by a data set which is
added to the spec? Should only be set to FALSE in special cases.}

\item{domain}{Either a vector with static values for the domain, or
a reactive that returns a such a vector.}
}
\description{
Add arbitrary scales to ggvis.
}
\keyword{internal}

