% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bin.R
\name{bin_vector}
\alias{bin_vector}
\alias{bin_vector.numeric}
\title{Bin vectors}
\usage{
bin_vector(x, weight = NULL, ...)

\method{bin_vector}{numeric}(
  x,
  weight = NULL,
  ...,
  width = 1,
  origin = 0,
  closed = c("right", "left"),
  pad = FALSE
)
}
\arguments{
\item{x}{A vector to bin}

\item{weight}{If specified, an integer vector of the same length as \code{x}
representing the number of occurances of each value in \code{x}}

\item{...}{additional arguments passed through to methods.}

\item{width}{The width of a bin}

\item{origin}{The left-most value for bins.}

\item{closed}{One of \code{"right"} or \code{"left"} indicating whether
right or left edges of bins are included in the bin.}

\item{pad}{A logical indicating whether the bins should be padded to include
an empty bin on each side.}
}
\description{
A generic and several implementations for binning vectors.
}
\keyword{internal}
