% Generated by roxygen2 (4.0.0): do not edit by hand
\name{compute_stack}
\alias{compute_stack}
\title{Stack overlapping data.}
\usage{
compute_stack(x, stack_var = NULL, group_var = NULL, ...)
}
\arguments{
\item{x}{A data object}

\item{stack_var}{A string specifying the stacking variable.}

\item{group_var}{A string specifying the grouping variable.}
}
\description{
Stack overlapping data.
}
\examples{
mtcars \%>\% cbind(count = 1) \%>\% compute_stack(count, cyl)

# Shouldn't use or affect existing grouping
mtcars \%>\% cbind(count = 1) \%>\% group_by(am) \%>\% compute_stack(count, cyl)

If given a ggvis object, will use x variable for stacking by default
mtcars \%>\% ggvis(x = ~cyl, y = ~wt) \%>\%
  compute_stack() \%>\%
  mark_rect(props(x = ~cyl - 0.5, x2 = ~cyl + 0.5,
                  y = ~stack_upr_, y2 = ~stack_lwr_))
}

