% Generated by roxygen2 (4.0.0): do not edit by hand
\name{datasource}
\alias{datasource}
\title{Create a new data source.}
\usage{
datasource(data, name = deparse2(substitute(data)))
}
\arguments{
\item{data}{a data frame}

\item{name}{the name of the data frame (used in error messages etc.)}
}
\description{
A data source implements a pointer (using a environment and a promise)
to an existing data frame or other data source in R. Most of the time you
should not need to call this function directly - just supply an object
to \code{\link{ggvis}} and \code{datasource} will be automatically called.
You may want to call this function directly in order to control the
datasource name used when printing and in the vega plot specification.
}
\examples{
datasource(mtcars)

# A simple example of a reactive data source
library(shiny)
v <- reactiveValues(n = 10)
p <- pipeline(reactive(mtcars[1:v$n, ]))
props <- props(x = ~wt, y = ~mpg)

sluice(p, props)

v$n <- 5
sluice(p, props)
}
\keyword{internal}

