% Generated by roxygen2 (4.0.0): do not edit by hand
\name{groups.ggvis}
\alias{arrange.ggvis}
\alias{dplyr-ggvis}
\alias{filter.ggvis}
\alias{groups.ggvis}
\alias{mutate.ggvis}
\alias{regroup.ggvis}
\alias{select.ggvis}
\alias{summarise.ggvis}
\alias{ungroup.ggvis}
\title{Dplyr verbs for ggvis}
\usage{
groups.ggvis(x)

regroup.ggvis(x, value)

ungroup.ggvis(x)

summarise.ggvis(.data, ...)

mutate.ggvis(.data, ...)

arrange.ggvis(.data, ...)

select.ggvis(.data, ...)

filter.ggvis(.data, ...)
}
\description{
Dplyr verbs for ggvis
}
\keyword{internal}

