% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Explain details of an object}
\description{
This is a generic function which gives more details about an object than
print, and is more focussed on human readable output than str.
}
\examples{
p <- mtcars \%>\% ggvis(x = ~cyl) \%>\% layer_bars()
explain(p)
}
\seealso{
\code{dplyr::\link[dplyr]{explain}} for more information.
}

