% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_png}
\alias{export_png}
\alias{export_svg}
\title{Export a PNG or SVG from a ggvis object}
\usage{
export_png(vis, file = NULL)

export_svg(vis, file = NULL)
}
\arguments{
\item{vis}{A ggvis object.}

\item{file}{Output file name. If NULL, defaults to "plot.svg" or "plot.png".}
}
\description{
This requires that the external program \code{vg2png} is installed. This is
part of the \code{vega} node.js module.
}
\examples{
\dontrun{
mtcars \%>\% ggvis(x = ~wt) \%>\% export_png()
}
}
\seealso{
\url{https://github.com/trifacta/vega} for information on installing
  \code{vg2png} and \code{vg2svg}.
}

