% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filter}
\alias{filter}
\title{Filter}
\arguments{
\item{.data}{A tbl.}

\item{...}{variables interpreted in the context of the data.}
}
\description{
This is the same as \code{dplyr::\link[dplyr]{filter}} function.
See \code{dplyr::\link[dplyr]{filter}} for more information.
}

