% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fullseq}
\alias{fullseq}
\title{Generate sequence of fixed size intervals covering range.}
\usage{
fullseq(range, size, ...)
}
\arguments{
\item{range}{range}

\item{size}{interval size}

\item{...}{other arguments passed on to methods}
}
\description{
Generate sequence of fixed size intervals covering range.
}
\seealso{
\code{\link[plyr]{round_any}}
}
\keyword{internal}

