% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dash.R
\name{geom_dash}
\alias{geom_dash}
\title{Line segments and curves}
\usage{
geom_dash(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., arrow = NULL, arrow.fill = NULL,
  lineend = "butt", linejoin = "round", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{arrow}{specification for arrow heads, as created by arrow().}

\item{arrow.fill}{fill colour to use for the arrow head (if closed). `NULL`
means use `colour` aesthetic.}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}
}
\description{
`geom_dash` draws a dash around points (x, y).
Couyld be useful for custom data visualisation,
or distribution or vector fields. For those scenarios it's more convenient than geom-segment.
}
\examples{

iris \%>\% ggplot(aes(Species , Petal.Length)) + geom_dash(width = 0.8)
}
