% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_data}
\alias{get_data}
\title{Get data from a ggvis object}
\usage{
get_data(vis)
}
\arguments{
\item{vis}{A ggvis object.}
}
\description{
This function is useful for inspecting the data in a ggvis object.
}
\examples{
p <- cocaine \%>\% ggvis(~price) \%>\% layer_bars()
get_data(p)
}

