% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ggvis}
\alias{ggvis}
\title{Visualise a data set with a ggvis graphic.}
\usage{
ggvis(data = NULL, ..., env = parent.frame())
}
\arguments{
\item{data}{A data object.}

\item{...}{Property mappings.}

\item{env}{Environment in which to evaluate properties.}
}
\description{
Visualise a data set with a ggvis graphic.
}
\examples{
layer_points(ggvis(mtcars, ~mpg, ~wt))
mtcars \%>\% ggvis(~mpg, ~wt) \%>\% layer_points()
}

