% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ggvis}
\alias{ggvis}
\title{Create a ggvis graphic.}
\usage{
ggvis(...)
}
\arguments{
\item{...}{components: data, \code{\link{props}}, \code{\link{layer}}es,
\code{\link{marks}}, \code{\link{scales}}, \code{\link{axis}} or
\code{\link{legend}} objects.}
}
\value{
a \code{layer} object. Will display the plot when printed;
  see \code{\link{save_spec}}, \code{\link{view_static}} and
  \code{\link{view_dynamic}} for other options.
}
\description{
Create a ggvis graphic.
}

