% Generated by roxygen2 (4.0.0): do not edit by hand
\name{handle_resize}
\alias{handle_resize}
\title{An interactive input bound to resize events.}
\usage{
handle_resize(vis, on_resize = NULL)
}
\arguments{
\item{vis}{Visualisation to listen to.}

\item{on_resize}{Callback function called with arguments \code{value} and
\code{session} every time plot size changes.}
}
\description{
An interactive input bound to resize events.
}
\examples{
# This example just prints out the current dimensions to the console
mtcars \%>\% ggvis(~mpg, ~wt) \%>\%
  layer_points() \%>\%
  handle_resize(function(value, ...) str(value))
}

