\docType{class}
\name{Hover-class}
\alias{Hover-class}
\title{Event broker for hover events.}
\description{
The hover event broker is useful if you want your shiny app
to respond to hover events (\code{mouse_out} and
\code{mouse_over}) in a custom way.
}
\seealso{
\code{\link{tooltip}} for a custom wrapper that uses hover
events to display tooltips.
}

