% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{is.dynamic}
\alias{is.dynamic}
\title{Determine if an ggvis is dynamic (i.e. needs to be run in a shiny app)}
\usage{
is.dynamic(x)
}
\description{
Determine if an ggvis is dynamic (i.e. needs to be run in a shiny app)
}
\keyword{internal}

