% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_is.R
\name{is.vega_title}
\alias{is.vega_title}
\title{Validate if the object is a vega title object}
\usage{
is.vega_title(obj, error = FALSE)
}
\arguments{
\item{obj}{an object}

\item{error}{Throw an error on parse failure? If TRUE, then the function returns NULL
on success (i.e., call only for the side-effect of an error on failure, like
stopifnot).}
}
\value{
logical
}
\description{
Validate if the object is a vega title object
}
