% Generated by roxygen2 (4.0.0): do not edit by hand
\name{linked_brush}
\alias{linked_brush}
\title{Create a linked brush object}
\usage{
linked_brush(keys = NULL, fill = "red")
}
\arguments{
\item{keys}{vector of all possible keys, if known.}

\item{fill}{brush colour}
}
\description{
Create a linked brush object
}

