% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark}
\alias{mark}
\alias{is.mark}
\title{Create a new "mark" object.}
\usage{
mark(type, props, data)

is.mark(x)
}
\arguments{
\item{type}{A string with the vega type.}

\item{props}{A list of properties, created by \code{\link{props}}.}

\item{data}{A reactive data object.}
}
\description{
A mark object is a close mapping to a vega mark object. Vega marks
are documented in \url{https://github.com/trifacta/vega/wiki/Marks}.
}
\details{
This function is designed to be used by authors of new types of mark.
}
\keyword{internal}
