% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%>\%}
\alias{\%>\%}
\title{Pipe graphics}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs,rhs}{A visualisation and a function to apply to it}
}
\description{
Like dplyr, ggvis also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\examples{
# Instead of
layer_points(ggvis(mtcars, ~mpg, ~wt))
# you can write
mtcars \%>\% ggvis(~mpg, ~wt) \%>\% layer_points()
}

