% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe graphics}
\arguments{
\item{lhs, rhs}{A visualisation and a function to apply to it}
}
\description{
Like dplyr, ggvis also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\examples{
# Instead of
layer_points(ggvis(mtcars, ~mpg, ~wt))
# you can write
mtcars \%>\% ggvis(~mpg, ~wt) \%>\% layer_points()
}
