\name{prop_value}
\alias{prop_value}
\title{Given a property and a dataset, get the value of the property.}
\usage{
  prop_value(x, data, processed)
}
\arguments{
  \item{x}{a vega property}

  \item{data}{a object containing data}

  \item{processed}{Has this data object been processed so
  that new columns have been calculated and unused columns
  have been dropped?}
}
\description{
  Given a property and a dataset, get the value of the
  property.
}

