% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvis.R
\name{save_spec}
\alias{save_spec}
\alias{view_spec}
\title{Tools to save and view static specs.}
\usage{
save_spec(x, path, ...)

view_spec(path, ...)
}
\arguments{
\item{x}{a ggvis object}

\item{path}{location to save spec to, or load spec from}

\item{...}{other arguments passed to \code{as.vega}}
}
\description{
These functions are mainly useful for testing.
}
\keyword{internal}

