\name{save_spec}
\alias{save_spec}
\alias{view_spec}
\title{Tools to save and view static specs.}
\usage{
  save_spec(path, x = last_vis(), ...)

  view_spec(path, ...)
}
\arguments{
  \item{path}{location to save spec to, or load spec from}

  \item{x}{a ggvis object}

  \item{...}{other arguments passed to \code{as.vega}}
}
\description{
  These functions are mainly useful for testing.
}
\keyword{internal}

