% Generated by roxygen2 (4.0.0): do not edit by hand
\name{scales}
\alias{is.scales}
\alias{scales}
\title{Create a "scales" object.}
\usage{
scales(..., .scales = list())

is.scales(x)
}
\arguments{
\item{...,.scales}{scales to combine into a single scales object}

\item{x}{object to test for scales-ness}
}
\description{
A scales object is used to manage multiple scales, essentially converting
a unnamed list into a named list.
}
\keyword{internal}

