% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interact_tooltip.R
\name{show_tooltip}
\alias{hide_tooltip}
\alias{show_tooltip}
\title{Send a message to the client to show or hide a tooltip}
\usage{
show_tooltip(session, l = 0, t = 0, html = "")

hide_tooltip(session)
}
\arguments{
\item{session}{A Shiny session object.}

\item{l}{Pixel location of left edge of tooltip (relative to page)}

\item{t}{Pixel location of top edge of tooltip (relative to page)}

\item{html}{HTML to display in the tooltip box.}
}
\description{
Send a message to the client to show or hide a tooltip
}

