% Generated by roxygen2 (4.0.1): do not edit by hand
\name{singular}
\alias{scale_singular}
\alias{singular}
\title{singular.}
\usage{
singular()

scale_singular(vis, scale, points = TRUE)
}
\arguments{
\item{vis}{A ggvis object.}

\item{points}{If \code{TRUE}, distributes the ordinal values over a
quantitative range at uniformly spaced points. The spacing of the points
can be adjusted using the padding property. If \code{FALSE}, the ordinal
scale will construct evenly-spaced bands, rather than points.}
}
\description{
Use singular when you want constant x or y position.
}
\examples{
mtcars \%>\% ggvis("", ~mpg) \%>\%
  layer_points() \%>\%
  scale_nominal("x") \%>\%
  add_guide_axis("x", title = "", tick_size_major = 0)

# OR
mtcars \%>\% ggvis("", ~mpg) \%>\%
  layer_points() \%>\%
  scale_singular("x")

# OR, even simpler
mtcars \%>\% ggvis(singular(), ~mpg) \%>\% layer_points()

# In the other direction:
mtcars \%>\% ggvis(~mpg, singular()) \%>\% layer_points()
}

