\name{source_reactive}
\alias{source_reactive}
\title{A reactive source.}
\usage{
  source_reactive(data, name = NULL)
}
\arguments{
  \item{data}{A reactive values object or reactive
  expression.}

  \item{name}{A name for the object.}
}
\description{
  A reactive source.
}
\examples{
source_reactive(reactive())
}

