\name{split_df}
\alias{is.split_df}
\alias{split_df}
\title{Create a data frame split up by specified variables}
\usage{
  split_df(data, split)

  is.split_df(x)
}
\arguments{
  \item{data}{a data frame}

  \item{split}{a list of properties}

  \item{x}{object to test for split_df-ness}
}
\description{
  Create a data frame split up by specified variables
}
\keyword{internal}

