% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subvis}
\alias{subvis}
\title{Create a subvisualisation.}
\usage{
subvis(vis, ..., data = NULL)
}
\description{
Create a subvisualisation.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

small <- nasaweather::atmos \%>\%
  filter(lat <= -11.217391, long <= -106.287, year == 1995) \%>\%
  group_by(long, lat)
small \%>\%
  ggvis(~long, ~lat) \%>\%
  layer_points()

small \%>\%
  ggvis(~long, ~lat) \%>\%
  subvis(width := 20, height := 20, stroke := "red") \%>\%
    layer_points(
      prop("x", ~month, scale = "x2"),
      prop("y", ~ozone, scale = "y2"))
}

