% Generated by roxygen2 (4.0.2): do not edit by hand
\name{subvis}
\alias{subvis}
\title{Create a subvisualisation.}
\usage{
subvis(vis, ..., data = NULL, width = NULL, height = NULL)
}
\description{
Create a subvisualisation.
}
\examples{
# Examples don't work yet
\dontrun{
library(dplyr, warn.conflicts = FALSE)

small <- nasaweather::atmos \%>\%
  filter(lat <= -11.217391, long <= -106.287, year == 1995) \%>\%
  group_by(long, lat)
small \%>\%
  ggvis(~long, ~lat) \%>\%
  layer_points()

small \%>\%
  ggvis(~long, ~lat) \%>\%
  subvis(width := 100, height := 100, stroke := "red") \%>\%
    layer_points(~month, ~ozone)

small \%>\%
  ggvis(~long, ~lat) \%>\%
  subvis(width := 100, height := 100, stroke := "red") \%>\%
    layer_points(~month, ~ozone) \%>\%
    add_axis("x", ticks = 3) \%>\%
    add_axis("y", ticks = 3)
}
}
\keyword{internal}

