\name{transform}
\alias{transform}
\title{S3 class: transform}
\usage{
  transform(type, ..., dots = list())
}
\arguments{
  \item{type}{A string representing type of transform.}

  \item{dots}{A list of arguments to pass to the underlying
  statistical transformation function.}

  \item{...}{Other arguments to pass to the specific
  transform.}
}
\description{
  This is a type of \code{\link{pipe}}.
}
\keyword{internal}

