% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_transform_filter.R
\name{vega_filter_transform}
\alias{vega_filter_transform}
\title{create a vega filter transform object}
\usage{
vega_filter_transform(type, expr = NULL)
}
\arguments{
\item{type}{All transforms require a type property, specifying the name of the transform.}

\item{expr}{Expr (required) A predicate [expression](../../expressions) for filtering the data. If the expression evaluates to `false`, the data object will be filtered.}
}
\value{
a transform object
}
\description{
https://vega.github.io/vega/docs/transforms/filter/index.html
NOTE: Some parameters are required to be 'arrays'.  In the case where a parameter must be an array but is a single line, wrap it in 'I()', (asis), to ensure it is properly handled.
}
\details{
The **filter** transform removes objects from a data stream based on a provided filter expression.
}
