% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega_refs_scale.R
\name{vega_scale}
\alias{vega_scale}
\title{create a vega scale object}
\usage{
vega_scale(name = NULL, type = "linear", align = NULL, base = NULL,
  clamp = NULL, domain = NULL, domainMax = NULL, domainMin = NULL,
  domainMid = NULL, domainRaw = NULL, exponent = NULL,
  interpolate = NULL, nice = NULL, padding = NULL, paddingInner = NULL,
  paddingOuter = NULL, range = NULL, reverse = NULL, round = NULL,
  zero = NULL)
}
\arguments{
\item{name}{String  Required. A unique name for the scale. Scales and projections share the same namespace; names must be unique across both.}

\item{type}{String  The type of scale (default linear). See the scale type reference for more.}

\item{align}{Number  The alignment of elements within each band step, as a fraction of the step size (default 0.5). This value must lie in the range \link{0,1}.}

\item{base}{Number  The base of the logarithm (default 10 for log scales).}

\item{clamp}{Boolean A boolean indicating if output values should be clamped to the range (default false). If clamping is disabled and the scale is passed a value outside the domain, the scale may return a value outside the range through extrapolation. If clamping is enabled, the output value of the scale is always within the scale’s range.}

\item{domain}{Domain  The domain of input data values for the scale. For quantitative data, this can take the form of a two-element array with minimum and maximum values. For ordinal or categorical data, this may be an array of valid input values. The domain may also be specified as a reference to a data source. See the scale domain reference for more.}

\item{domainMax}{Number  Sets the maximum value in the scale domain, overriding the domain property. The domainMax property is only intended for use with scales having continuous domains.}

\item{domainMin}{Number  Sets the minimum value in the scale domain, overriding the domain property. The domainMin property is only intended for use with scales having continuous domains.}

\item{domainMid}{Number  Inserts a single mid-point value into a two-element domain. The mid-point value must lie between the domain minium and maximum values. This property can be useful for setting a midpoint for diverging color scales. The domainMid property is only intended for use with scales supporting continuous, piecewise domains.}

\item{domainRaw}{Array An array of raw values that, if non-null, directly overrides the domain property. This is useful for supporting interactions such as panning or zooming a scale. The scale may be initially determined using a data-driven domain, then modified in response to user input by setting the rawDomain value.}

\item{exponent}{Number  The exponent to use in the scale transform (default 1 for power scales).}

\item{interpolate}{String | Object The interpolation method for range values. By default, a general interpolator for numbers, dates, strings and colors (in RGB space) is used. For color ranges, this property allows interpolation in alternative color spaces. Legal values include rgb, hsl, hsl-long, lab, hcl, hcl-long, cubehelix and cubehelix-long (‘-long’ variants use longer paths in polar coordinate spaces). If object-valued, this property accepts an object with a string-valued type property and an optional numeric gamma property applicable to rgb and cubehelix interpolators. For more, see the d3-interpolate documentation.}

\item{nice}{String | Object | Number | Boolean  Multiple definitions based on scale type. See scale type documentation.}

\item{padding}{Number Multiple definitions based on scale type. See scale type documentation.}

\item{paddingInner}{Number  The inner padding (spacing) within each band step, as a fraction of the step size (default 0). This value must lie in the range \link{0,1}.}

\item{paddingOuter}{Number  The outer padding (spacing) at the ends of the scale range, as a fraction of the step size (default 0). This value must lie in the range \link{0,1}.}

\item{range}{| Scheme | Color[] Range The range of the scale, representing the set of visual values. For numeric values, the range can take the form of a two-element array with minimum and maximum values. For ordinal or quantized data, the range may be an array of desired output values, which are mapped to elements in the specified domain. See the scale range reference for more.}

\item{reverse}{Boolean If true, reverses the order of the scale range.}

\item{round}{Boolean If true, rounds numeric output values to integers. Helpful for snapping to the pixel grid.}

\item{zero}{Boolean Boolean flag indicating if the scale domain should include zero. The default value is true for linear, sqrt and pow, and false otherwise.}
}
\value{
a vega scale object
}
\description{
https://vega.github.io/vega/docs/scales/
}
