\name{view_static}
\alias{view_static}
\title{Generate a static web page with the embedded ggvis graph}
\usage{
  view_static(gv, renderer = "canvas", launch = TRUE)
}
\arguments{
  \item{gv}{A ggvis object.}

  \item{renderer}{The renderer to use in the browser. Can
  be \code{"canvas"} (the default) or \code{"svg"}.}

  \item{launch}{If \code{TRUE}, launch this web page in a
  browser.}
}
\description{
  Generate a static web page with the embedded ggvis graph
}
\keyword{internal}

